Reorganize code lines
