--- Progetto/regi_ui.py
@@ -1,2012 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Form implementation generated from reading ui file 'regi_ui.ui'
-#
-# Created by: PyQt5 UI code generator 5.13.2
-#
-# WARNING! All changes made in this file will be lost!
-
-
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-import sentry_sdk
-sentry_sdk.init("https://d770165f943e497484a6fa95ebcef724@o429982.ingest.sentry.io/5377715")
-from regi_db import db, Value, Cliente, Preventivo, TestoIn, TestoFin, Progetto, Infisso
-from regi_db import model_to_dict
-from datetime import date
-
-class TreeItem(QtGui.QStandardItem):
-    def __init__(self, txt = "", font_size = 12, set_bold = False, color = QtGui.QColor(0, 0, 0)):
-        super().__init__()
-        self.setEditable(False)
-        fnt = QtGui.QFont("Open Sans", font_size)
-        fnt.setBold(set_bold)
-        self.setFont(fnt)
-        self.setText(txt)
-        self.setForeground(color)
-
-class DialogAggCliente(QtWidgets.QDialog): 
-    def __init__(self, parent=None):
-        super(QtWidgets.QDialog, self).__init__(parent)
-        self.resize(600, 200)
-        self.edit_cognome = QtWidgets.QLineEdit()
-        self.edit_nome = QtWidgets.QLineEdit()
-        self.edit_via = QtWidgets.QLineEdit()
-        self.edit_comune = QtWidgets.QLineEdit()
-        self.edit_tel1 = QtWidgets.QLineEdit()
-        self.edit_tel2 = QtWidgets.QLineEdit()
-        self.edit_email = QtWidgets.QLineEdit()
-        self.edit_cantiere = QtWidgets.QLineEdit()
-        layout_agg_cliente = QtWidgets.QFormLayout()
-        layout_agg_cliente.addRow("Cognome", self.edit_cognome)
-        layout_agg_cliente.addRow("Nome", self.edit_nome)
-        layout_agg_cliente.addRow("Via", self.edit_via)
-        layout_agg_cliente.addRow("Comune", self.edit_comune)
-        layout_agg_cliente.addRow("Telefono 1", self.edit_tel1)
-        layout_agg_cliente.addRow("Telefono 2", self.edit_tel2)
-        layout_agg_cliente.addRow("Email", self.edit_email)
-        layout_agg_cliente.addRow("Cantiere", self.edit_cantiere)
-        self.bottoni = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel)
-        self.bottoni.accepted.connect(self.controlla)
-        self.bottoni.rejected.connect(self.reject)
-        layout = QtWidgets.QVBoxLayout()
-        layout.addLayout(layout_agg_cliente)
-        layout.addWidget(self.bottoni)
-        self.setLayout(layout)
-    def popup_nome_mancante(self):
-        popup = QtWidgets.QMessageBox()
-        popup.resize(600, 300)
-        popup.setWindowTitle("Parametri non corretti")
-        popup.setText("Per caricare i dati nell'archivio sono necessari almeno i seguenti dati: Cognome, Nome. Si prega di riempire questi campi.")
-        popup.setIcon(QtWidgets.QMessageBox.Information)
-        x = popup.exec_()
-    def controlla(self):
-        if self.edit_cognome.text() == "" or self.edit_nome.text() == "":
-            self.popup_nome_mancante()
-        else:
-            self.cliente_nuovo = {"cognome": self.edit_cognome.text(),
-                            "nome": self.edit_nome.text(),
-                            "via": self.edit_via.text(),
-                            "comune": self.edit_comune.text(),
-                            "tel1": self.edit_tel1.text(),
-                            "tel2": self.edit_tel2.text(),
-                            "email": self.edit_email.text(),
-                            "cantiere": self.edit_cantiere.text()}
-            self.accept()
-
-class DialogEditCliente(QtWidgets.QDialog): 
-    def __init__(self, parent=None):
-        super(QtWidgets.QDialog, self).__init__(parent)
-        self.resize(600, 200)
-        self.edit_cognome = QtWidgets.QLineEdit()
-        self.edit_nome = QtWidgets.QLineEdit()
-        self.edit_via = QtWidgets.QLineEdit()
-        self.edit_comune = QtWidgets.QLineEdit()
-        self.edit_tel1 = QtWidgets.QLineEdit()
-        self.edit_tel2 = QtWidgets.QLineEdit()
-        self.edit_email = QtWidgets.QLineEdit()
-        self.edit_cantiere = QtWidgets.QLineEdit()
-        self.edit_cognome.setText(ui.current_cliente["cognome"])
-        self.edit_nome.setText(ui.current_cliente["nome"])
-        self.edit_via.setText(ui.current_cliente["via"])
-        self.edit_comune.setText(ui.current_cliente["comune"])
-        self.edit_tel1.setText(ui.current_cliente["tel1"])
-        self.edit_tel2.setText(ui.current_cliente["tel2"])
-        self.edit_email.setText(ui.current_cliente["email"])
-        self.edit_cantiere.setText(ui.current_cliente["cantiere"])
-        layout_edit_cliente = QtWidgets.QFormLayout()
-        layout_edit_cliente.addRow("Cognome", self.edit_cognome)
-        layout_edit_cliente.addRow("Nome", self.edit_nome)
-        layout_edit_cliente.addRow("Via", self.edit_via)
-        layout_edit_cliente.addRow("Comune", self.edit_comune)
-        layout_edit_cliente.addRow("Telefono 1", self.edit_tel1)
-        layout_edit_cliente.addRow("Telefono 2", self.edit_tel2)
-        layout_edit_cliente.addRow("Email", self.edit_email)
-        layout_edit_cliente.addRow("Cantiere", self.edit_cantiere)
-        self.bottoni = QtWidgets.QDialogButtonBox(QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel)
-        self.bottoni.accepted.connect(self.controlla)
-        self.bottoni.rejected.connect(self.reject)
-        layout = QtWidgets.QVBoxLayout()
-        layout.addLayout(layout_edit_cliente)
-        layout.addWidget(self.bottoni)
-        self.setLayout(layout)
-    def popup_nome_mancante(self):
-        popup = QtWidgets.QMessageBox()
-        popup.resize(600, 300)
-        popup.setWindowTitle("Parametri non corretti")
-        popup.setText("Per caricare i dati nell'archivio sono necessari almeno i seguenti dati: Cognome, Nome. Si prega di riempire questi campi.")
-        popup.setIcon(QtWidgets.QMessageBox.Information)
-        x = popup.exec_()
-    def controlla(self):
-        if self.edit_cognome.text() == "" or self.edit_nome.text() == "":
-            self.popup_nome_mancante()
-        else:
-            self.cliente_modificato = {"cognome": self.edit_cognome.text(),
-                            "nome": self.edit_nome.text(),
-                            "via": self.edit_via.text(),
-                            "comune": self.edit_comune.text(),
-                            "tel1": self.edit_tel1.text(),
-                            "tel2": self.edit_tel2.text(),
-                            "email": self.edit_email.text(),
-                            "cantiere": self.edit_cantiere.text()}
-            self.accept()
-
-class DialogScegliTesto(QtWidgets.QDialog): 
-    def __init__(self, parent = None):
-        super(QtWidgets.QDialog, self).__init__(parent)
-        self.setWindowTitle("Scegli il testo iniziale/finale")
-        self.resize(700, 500)
-        self.btn_aggiungi = QtWidgets.QPushButton()
-        self.stack.addWidget(self.btn_aggiungi)
-        self.btn_aggiungi.setGeometry(QtCore.QRect(10, 450, 80, 41))
-        self.btn_aggiungi.setText("Aggiungi")
-        self.btn_modifica = QtWidgets.QPushButton()
-        self.btn_modifica.setGeometry(QtCore.QRect(100, 450, 80, 41))
-        self.btn_modifica.setText("btn_modifica")
-        self.btn_elimina = QtWidgets.QPushButton()
-        self.btn_elimina.setGeometry(QtCore.QRect(190, 450, 80, 41))
-        self.btn_elimina.setText("btn_elimina")
-        self.btn_scegli = QtWidgets.QPushButton()
-        self.btn_scegli.setGeometry(QtCore.QRect(610, 450, 80, 41))
-        self.btn_scegli.setText("btn_scegli")
-        self.edit_testo_selez = QtWidgets.QTextEdit()
-        self.edit_testo_selez.setEnabled(True)
-        self.edit_testo_selez.setGeometry(QtCore.QRect(10, 320, 581, 111))
-        self.edit_testo_selez.setFrameShape(QtWidgets.QFrame.NoFrame)
-        self.edit_testo_selez.setReadOnly(True)
-        self.tab_testi = QtWidgets.QTableWidget()
-        self.tab_testi.setGeometry(QtCore.QRect(10, 10, 681, 301))
-        self.tab_testi.setColumnCount(2)
-        self.tab_testi.setRowCount(0)
-        self.tab_testi.setHorizontalHeaderLabels(["ID", "Testo"])
-        self.tab_testi.horizontalHeader().setStretchLastSection(True)
-        self.btn_fatto = QtWidgets.QPushButton()
-        self.btn_fatto.setGeometry(QtCore.QRect(610, 350, 80, 41))
-        self.btn_fatto.setText("Fatto")
-
-        # parte del layout mancante, finestra per ora invisibile
-        # wip
-
-        self.operazione = ""
-        self.popola_tabella()
-        self.tab_testi.cellClicked.connect(self.popola_edit)
-        self.btn_aggiungi.clicked.connect(self.aggiungi)
-        self.btn_modifica.clicked.connect(self.modifica)
-        self.btn_elimina.clicked.connect(self.elimina)
-        self.btn_fatto.clicked.connect(self.fatto)
-        self.btn_scegli.clicked.connect(self.scegli)
-    
-    def popola_tabella(self):
-        if ui.quale_testo == "in":
-            lista_testi = TestoIn.select().dicts()
-        else:
-            lista_testi = TestoFin.select().dicts()
-        for num_riga, voce in enumerate(lista_testi):
-            self.tab_testi.insertRow(num_riga)
-            self.tab_testi.setItem(num_riga, 0, QtWidgets.QTableWidgetItem(str(voce["id"]).zfill(4)))
-            self.tab_testi.setItem(num_riga, 1, QtWidgets.QTableWidgetItem(voce["testo"][:100]))
-
-    def popola_edit(self):
-        if self.tab_testi.currentRow() == -1:
-            self.popup_testo("Nessun testo selezionato")
-        elif self.operazione == "":
-            id_testo_selez = self.tab_testi.item(self.tab_testi.currentRow(), 0).text()
-            if ui.quale_testo == "in":
-                voce_selez = TestoIn.get(TestoIn.id == id_testo_selez)
-            else:
-                voce_selez = TestoFin.get(TestoFin.id == id_testo_selez)
-            self.edit_testo_selez.setPlainText(voce_selez.testo)
-
-    def popup_testo(self, titolo):
-        popup = QtWidgets.QMessageBox()
-        popup.resize(600, 300)
-        popup.setWindowTitle(titolo)
-        if titolo == "Nessun testo selezionato":
-            popup.setText("Selezionare un testo per modificarlo o eliminarlo.")
-        elif titolo == "Testo vuoto":
-            popup.setText("Il contenuto del testo non può essere vuoto.")
-        popup.setIcon(QtWidgets.QMessageBox.Information)
-        x = popup.exec_()
-
-    def aggiungi(self):
-        self.operazione = "aggiungi"
-        self.btn_modifica.setEnabled(False)
-        self.btn_elimina.setEnabled(False)
-        self.btn_scegli.setEnabled(False)
-        self.edit_testo_selez.setPlainText("")
-        self.edit_testo_selez.setReadOnly(False)
-        self.edit_testo_selez.setFrameShape(QtWidgets.QFrame.StyledPanel)
-
-
-    def modifica(self):
-        self.operazione = "modifica"
-        self.btn_aggiungi.setEnabled(False)
-        self.btn_elimina.setEnabled(False)
-        self.btn_scegli.setEnabled(False)
-        self.edit_testo_selez.setFrameShape(QtWidgets.QFrame.StyledPanel)
-
-    def elimina(self):
-        if self.tab_testi.currentRow() == -1:
-            self.popup_testo("Nessun testo selezionato")
-        else:
-            id_testo_canc = self.tab_testi.item(self.tab_testi.currentRow(), 0).text()
-            if ui.quale_testo == "in":
-                TestoIn.get(TestoIn.id == id_testo_canc).delete_instance()
-            else:
-                TestoFin.get(TestoFin.id == id_testo_canc).delete_instance()
-            self.tab_testi.removeRow(self.tab_testi.currentRow())
-
-    def fatto(self):
-        if self.edit_testo_selez.toPlainText() == "":
-            self.popup_testo("Testo vuoto")
-        
-        elif self.operazione == "aggiungi":
-            if ui.quale_testo == "in":
-                t = model_to_dict(TestoIn.create(testo = self.edit_testo_selez.toPlainText()))
-            else:
-                t = model_to_dict(TestoFin.create(testo = self.edit_testo_selez.toPlainText()))
-            num_riga = self.tab_testi.rowCount()
-            self.tab_testi.insertRow(num_riga)
-            self.tab_testi.setItem(num_riga, 0, QtWidgets.QTableWidgetItem(str(t["id"]).zfill(4)))
-            self.tab_testi.setItem(num_riga, 1, QtWidgets.QTableWidgetItem(t["testo"]))
-            self.btn_modifica.setEnabled(True)
-            self.btn_elimina.setEnabled(True)
-            self.btn_scegli.setEnabled(True)
-
-        elif self.operazione == "modifica":
-            num_riga = self.tab_testi.currentRow()
-            id_testo_modificato = self.tab_testi.item(num_riga, 0).text()
-            testo_modificato = self.edit_testo_selez.toPlainText()
-            if ui.quale_testo == "in":
-                query = TestoIn.update(testo_modificato).where(TestoIn.id == id_testo_modificato)
-            else:
-                query = TestoFin.update(testo_modificato).where(TestoFin.id == id_testo_modificato)
-            query.execute()
-            self.tab_testi.setItem(num_riga, 1, QtWidgets.QTableWidgetItem(testo_modificato[:100]))
-            self.btn_aggiungi.setEnabled(True)
-            self.btn_elimina.setEnabled(True)
-            self.btn_scegli.setEnabled(True)
-
-        self.operazione = ""
-        self.edit_testo_selez.setReadOnly(True)
-        self.edit_testo_selez.setFrameShape(QtWidgets.QFrame.NoFrame)
-
-    def scegli(self):
-        self.testo_da_passare = self.edit_testo_selez.toPlainText()
-        self.accept()
-
-
-class Ui_reGI(object):
-    def setupUi(self, reGI):
-        reGI.setObjectName("reGI")
-        reGI.resize(1920, 1010)
-        icon = QtGui.QIcon()
-        icon.addPixmap(QtGui.QPixmap("../../../.designer/Icone/regi_1.0.png"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
-        reGI.setWindowIcon(icon)
-        reGI.setAutoFillBackground(True)
-        reGI.setIconSize(QtCore.QSize(30, 30))
-        reGI.setTabShape(QtWidgets.QTabWidget.Rounded)
-        self.centralWidget = QtWidgets.QWidget(reGI)
-        self.centralWidget.setAutoFillBackground(True)
-        self.centralWidget.setObjectName("centralWidget")
-        self.stackedWidget = QtWidgets.QStackedWidget(self.centralWidget)
-        self.stackedWidget.setGeometry(QtCore.QRect(10, 10, 1901, 951))
-        self.stackedWidget.setAutoFillBackground(True)
-        self.stackedWidget.setObjectName("stackedWidget")
-        self.pag_0 = QtWidgets.QWidget()
-        self.pag_0.setObjectName("pag_0")
-        self.horizontalLayoutWidget = QtWidgets.QWidget(self.pag_0)
-        self.horizontalLayoutWidget.setGeometry(QtCore.QRect(-1, 10, 1891, 941))
-        self.horizontalLayoutWidget.setObjectName("horizontalLayoutWidget")
-        self.horizontalLayout = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget)
-        self.horizontalLayout.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout.setObjectName("horizontalLayout")
-        self.groupBox = QtWidgets.QGroupBox(self.horizontalLayoutWidget)
-        font = QtGui.QFont()
-        font.setPointSize(14)
-        font.setBold(True)
-        font.setWeight(75)
-        self.groupBox.setFont(font)
-        self.groupBox.setFlat(False)
-        self.groupBox.setCheckable(False)
-        self.groupBox.setObjectName("groupBox")
-        self.label_3 = QtWidgets.QLabel(self.groupBox)
-        self.label_3.setGeometry(QtCore.QRect(30, 70, 51, 40))
-        font = QtGui.QFont()
-        font.setPointSize(14)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_3.setFont(font)
-        self.label_3.setObjectName("label_3")
-        self.label_4 = QtWidgets.QLabel(self.groupBox)
-        self.label_4.setGeometry(QtCore.QRect(30, 280, 71, 40))
-        font = QtGui.QFont()
-        font.setPointSize(14)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_4.setFont(font)
-        self.label_4.setObjectName("label_4")
-        self.btn_nuovo_prev = QtWidgets.QPushButton(self.groupBox)
-        self.btn_nuovo_prev.setEnabled(True)
-        self.btn_nuovo_prev.setGeometry(QtCore.QRect(30, 120, 175, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setItalic(False)
-        font.setWeight(75)
-        self.btn_nuovo_prev.setFont(font)
-        self.btn_nuovo_prev.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_nuovo_prev.setObjectName("btn_nuovo_prev")
-        self.btn_apri_prev = QtWidgets.QPushButton(self.groupBox)
-        self.btn_apri_prev.setGeometry(QtCore.QRect(30, 170, 175, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.btn_apri_prev.setFont(font)
-        self.btn_apri_prev.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_apri_prev.setObjectName("btn_apri_prev")
-        self.listView = QtWidgets.QListView(self.groupBox)
-        self.listView.setGeometry(QtCore.QRect(30, 330, 250, 250))
-        self.listView.setObjectName("listView")
-        self.horizontalLayout.addWidget(self.groupBox)
-        self.groupBox_2 = QtWidgets.QGroupBox(self.horizontalLayoutWidget)
-        font = QtGui.QFont()
-        font.setPointSize(14)
-        font.setBold(True)
-        font.setWeight(75)
-        self.groupBox_2.setFont(font)
-        self.groupBox_2.setObjectName("groupBox_2")
-        self.btn_view_anagrafica = QtWidgets.QPushButton(self.groupBox_2)
-        self.btn_view_anagrafica.setGeometry(QtCore.QRect(30, 120, 250, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.btn_view_anagrafica.setFont(font)
-        self.btn_view_anagrafica.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_view_anagrafica.setObjectName("btn_view_anagrafica")
-        self.btn_sc_agg_cliente = QtWidgets.QPushButton(self.groupBox_2)
-        self.btn_sc_agg_cliente.setGeometry(QtCore.QRect(30, 170, 250, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.btn_sc_agg_cliente.setFont(font)
-        self.btn_sc_agg_cliente.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_sc_agg_cliente.setObjectName("btn_sc_agg_cliente")
-        self.btn_sc_edit_cliente = QtWidgets.QPushButton(self.groupBox_2)
-        self.btn_sc_edit_cliente.setGeometry(QtCore.QRect(30, 220, 250, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.btn_sc_edit_cliente.setFont(font)
-        self.btn_sc_edit_cliente.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_sc_edit_cliente.setObjectName("btn_sc_edit_cliente")
-        self.horizontalLayout.addWidget(self.groupBox_2)
-        self.groupBox_3 = QtWidgets.QGroupBox(self.horizontalLayoutWidget)
-        font = QtGui.QFont()
-        font.setPointSize(14)
-        font.setBold(True)
-        font.setWeight(75)
-        self.groupBox_3.setFont(font)
-        self.groupBox_3.setObjectName("groupBox_3")
-        self.btn_view_modelli = QtWidgets.QPushButton(self.groupBox_3)
-        self.btn_view_modelli.setGeometry(QtCore.QRect(30, 120, 250, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.btn_view_modelli.setFont(font)
-        self.btn_view_modelli.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_view_modelli.setObjectName("btn_view_modelli")
-        self.btn_agg_modello = QtWidgets.QPushButton(self.groupBox_3)
-        self.btn_agg_modello.setGeometry(QtCore.QRect(30, 170, 250, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.btn_agg_modello.setFont(font)
-        self.btn_agg_modello.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_agg_modello.setObjectName("btn_agg_modello")
-        self.btn_edit_modello = QtWidgets.QPushButton(self.groupBox_3)
-        self.btn_edit_modello.setGeometry(QtCore.QRect(30, 220, 250, 30))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.btn_edit_modello.setFont(font)
-        self.btn_edit_modello.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_edit_modello.setObjectName("btn_edit_modello")
-        self.horizontalLayout.addWidget(self.groupBox_3)
-        self.stackedWidget.addWidget(self.pag_0)
-        self.pag_1 = QtWidgets.QWidget()
-        self.pag_1.setObjectName("pag_1")
-        self.tab_clienti = QtWidgets.QTableWidget(self.pag_1)
-        self.tab_clienti.setEnabled(True)
-        self.tab_clienti.setGeometry(QtCore.QRect(10, 80, 1881, 691))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.tab_clienti.setFont(font)
-        self.tab_clienti.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.tab_clienti.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.tab_clienti.setSizeAdjustPolicy(QtWidgets.QAbstractScrollArea.AdjustToContents)
-        self.tab_clienti.setEditTriggers(QtWidgets.QAbstractItemView.NoEditTriggers)
-        self.tab_clienti.setAlternatingRowColors(True)
-        self.tab_clienti.setSelectionMode(QtWidgets.QAbstractItemView.SingleSelection)
-        self.tab_clienti.setSelectionBehavior(QtWidgets.QAbstractItemView.SelectRows)
-        self.tab_clienti.setIconSize(QtCore.QSize(0, 0))
-        self.tab_clienti.setGridStyle(QtCore.Qt.DotLine)
-        self.tab_clienti.setObjectName("tab_clienti")
-        self.tab_clienti.setColumnCount(9)
-        self.tab_clienti.setRowCount(0)
-        item = QtWidgets.QTableWidgetItem()
-        self.tab_clienti.setHorizontalHeaderItem(0, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tab_clienti.setHorizontalHeaderItem(1, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tab_clienti.setHorizontalHeaderItem(2, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tab_clienti.setHorizontalHeaderItem(3, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tab_clienti.setHorizontalHeaderItem(4, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tab_clienti.setHorizontalHeaderItem(5, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tab_clienti.setHorizontalHeaderItem(6, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tab_clienti.setHorizontalHeaderItem(7, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tab_clienti.setHorizontalHeaderItem(8, item)
-        self.tab_clienti.horizontalHeader().setCascadingSectionResizes(True)
-        self.tab_clienti.horizontalHeader().setDefaultSectionSize(200)
-        self.tab_clienti.verticalHeader().setVisible(False)
-        self.label_2 = QtWidgets.QLabel(self.pag_1)
-        self.label_2.setGeometry(QtCore.QRect(15, 30, 170, 30))
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_2.setFont(font)
-        self.label_2.setObjectName("label_2")
-        self.frame_2 = QtWidgets.QFrame(self.pag_1)
-        self.frame_2.setGeometry(QtCore.QRect(10, 800, 711, 70))
-        self.frame_2.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame_2.setFrameShadow(QtWidgets.QFrame.Plain)
-        self.frame_2.setObjectName("frame_2")
-        self.btn_edit_cliente = QtWidgets.QPushButton(self.frame_2)
-        self.btn_edit_cliente.setEnabled(True)
-        self.btn_edit_cliente.setGeometry(QtCore.QRect(370, 10, 150, 50))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.btn_edit_cliente.setFont(font)
-        self.btn_edit_cliente.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_edit_cliente.setObjectName("btn_edit_cliente")
-        self.btn_scegli_cliente = QtWidgets.QPushButton(self.frame_2)
-        self.btn_scegli_cliente.setGeometry(QtCore.QRect(10, 10, 150, 50))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.btn_scegli_cliente.setFont(font)
-        self.btn_scegli_cliente.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_scegli_cliente.setObjectName("btn_scegli_cliente")
-        self.btn_agg_cliente = QtWidgets.QPushButton(self.frame_2)
-        self.btn_agg_cliente.setEnabled(True)
-        self.btn_agg_cliente.setGeometry(QtCore.QRect(190, 10, 150, 50))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.btn_agg_cliente.setFont(font)
-        self.btn_agg_cliente.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_agg_cliente.setObjectName("btn_agg_cliente")
-        self.btn_canc_cliente = QtWidgets.QPushButton(self.frame_2)
-        self.btn_canc_cliente.setEnabled(True)
-        self.btn_canc_cliente.setGeometry(QtCore.QRect(550, 10, 150, 50))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.btn_canc_cliente.setFont(font)
-        self.btn_canc_cliente.setCursor(QtGui.QCursor(QtCore.Qt.PointingHandCursor))
-        self.btn_canc_cliente.setObjectName("btn_canc_cliente")
-        self.stackedWidget.addWidget(self.pag_1)
-        self.pag_2 = QtWidgets.QWidget()
-        self.pag_2.setObjectName("pag_2")
-        self.tab_dettaglio_totale_stampa = QtWidgets.QTabWidget(self.pag_2)
-        self.tab_dettaglio_totale_stampa.setGeometry(QtCore.QRect(10, 160, 1881, 791))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        font.setKerning(True)
-        self.tab_dettaglio_totale_stampa.setFont(font)
-        self.tab_dettaglio_totale_stampa.setTabShape(QtWidgets.QTabWidget.Rounded)
-        self.tab_dettaglio_totale_stampa.setIconSize(QtCore.QSize(16, 20))
-        self.tab_dettaglio_totale_stampa.setObjectName("tab_dettaglio_totale_stampa")
-        self.scheda_dettaglio = QtWidgets.QWidget()
-        self.scheda_dettaglio.setObjectName("scheda_dettaglio")
-        self.listwidget_inf = QtWidgets.QListWidget(self.scheda_dettaglio)
-        self.listwidget_inf.setGeometry(QtCore.QRect(10, 90, 390, 660))
-        font = QtGui.QFont()
-        font.setPointSize(12)
-        font.setBold(True)
-        font.setWeight(75)
-        self.listwidget_inf.setFont(font)
-        self.listwidget_inf.setProperty("showDropIndicator", False)
-        self.listwidget_inf.setAlternatingRowColors(True)
-        self.listwidget_inf.setSelectionBehavior(QtWidgets.QAbstractItemView.SelectItems)
-        self.listwidget_inf.setIconSize(QtCore.QSize(20, 20))
-        self.listwidget_inf.setObjectName("listwidget_inf")
-        item = QtWidgets.QListWidgetItem()
-        icon1 = QtGui.QIcon()
-        icon1.addPixmap(QtGui.QPixmap("../../../.designer/Icone/regi_0.0.png"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
-        item.setIcon(icon1)
-        self.listwidget_inf.addItem(item)
-        self.label_23 = QtWidgets.QLabel(self.scheda_dettaglio)
-        self.label_23.setGeometry(QtCore.QRect(433, 13, 121, 20))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_23.setFont(font)
-        self.label_23.setObjectName("label_23")
-        self.edit_cod_prog = QtWidgets.QLineEdit(self.scheda_dettaglio)
-        self.edit_cod_prog.setGeometry(QtCore.QRect(554, 10, 91, 25))
-        self.edit_cod_prog.setText("")
-        self.edit_cod_prog.setPlaceholderText("")
-        self.edit_cod_prog.setObjectName("edit_cod_prog")
-        self.label_24 = QtWidgets.QLabel(self.scheda_dettaglio)
-        self.label_24.setGeometry(QtCore.QRect(435, 450, 81, 16))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_24.setFont(font)
-        self.label_24.setObjectName("label_24")
-        self.edit_note_descr = QtWidgets.QLineEdit(self.scheda_dettaglio)
-        self.edit_note_descr.setGeometry(QtCore.QRect(520, 530, 550, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        font.setBold(False)
-        font.setWeight(50)
-        self.edit_note_descr.setFont(font)
-        self.edit_note_descr.setClearButtonEnabled(True)
-        self.edit_note_descr.setObjectName("edit_note_descr")
-        self.label_25 = QtWidgets.QLabel(self.scheda_dettaglio)
-        self.label_25.setGeometry(QtCore.QRect(440, 535, 41, 16))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_25.setFont(font)
-        self.label_25.setObjectName("label_25")
-        self.btn_scegli_prog = QtWidgets.QToolButton(self.scheda_dettaglio)
-        self.btn_scegli_prog.setGeometry(QtCore.QRect(650, 10, 25, 25))
-        self.btn_scegli_prog.setObjectName("btn_scegli_prog")
-        self.line_5 = QtWidgets.QFrame(self.scheda_dettaglio)
-        self.line_5.setGeometry(QtCore.QRect(405, 10, 20, 741))
-        self.line_5.setFrameShape(QtWidgets.QFrame.VLine)
-        self.line_5.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line_5.setObjectName("line_5")
-        self.edit_descr = QtWidgets.QTextEdit(self.scheda_dettaglio)
-        self.edit_descr.setGeometry(QtCore.QRect(520, 445, 521, 70))
-        font = QtGui.QFont()
-        font.setBold(True)
-        font.setWeight(75)
-        self.edit_descr.setFont(font)
-        self.edit_descr.setObjectName("edit_descr")
-        self.label_26 = QtWidgets.QLabel(self.scheda_dettaglio)
-        self.label_26.setEnabled(False)
-        self.label_26.setGeometry(QtCore.QRect(810, 570, 51, 16))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_26.setFont(font)
-        self.label_26.setObjectName("label_26")
-        self.checkbox_foto_3d = QtWidgets.QCheckBox(self.scheda_dettaglio)
-        self.checkbox_foto_3d.setEnabled(False)
-        self.checkbox_foto_3d.setGeometry(QtCore.QRect(810, 620, 91, 17))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.checkbox_foto_3d.setFont(font)
-        self.checkbox_foto_3d.setObjectName("checkbox_foto_3d")
-        self.checkbox_foto_2d = QtWidgets.QCheckBox(self.scheda_dettaglio)
-        self.checkbox_foto_2d.setEnabled(False)
-        self.checkbox_foto_2d.setGeometry(QtCore.QRect(810, 595, 91, 17))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.checkbox_foto_2d.setFont(font)
-        self.checkbox_foto_2d.setChecked(True)
-        self.checkbox_foto_2d.setTristate(False)
-        self.checkbox_foto_2d.setObjectName("checkbox_foto_2d")
-        self.label_27 = QtWidgets.QLabel(self.scheda_dettaglio)
-        self.label_27.setGeometry(QtCore.QRect(810, 13, 61, 20))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_27.setFont(font)
-        self.label_27.setObjectName("label_27")
-        self.edit_posiz = QtWidgets.QLineEdit(self.scheda_dettaglio)
-        self.edit_posiz.setGeometry(QtCore.QRect(874, 10, 196, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        font.setBold(False)
-        font.setWeight(50)
-        self.edit_posiz.setFont(font)
-        self.edit_posiz.setText("")
-        self.edit_posiz.setPlaceholderText("")
-        self.edit_posiz.setClearButtonEnabled(True)
-        self.edit_posiz.setObjectName("edit_posiz")
-        self.line_6 = QtWidgets.QFrame(self.scheda_dettaglio)
-        self.line_6.setGeometry(QtCore.QRect(1080, 10, 20, 741))
-        self.line_6.setFrameShape(QtWidgets.QFrame.VLine)
-        self.line_6.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line_6.setObjectName("line_6")
-        self.line_7 = QtWidgets.QFrame(self.scheda_dettaglio)
-        self.line_7.setGeometry(QtCore.QRect(1440, 10, 20, 741))
-        self.line_7.setFrameShape(QtWidgets.QFrame.VLine)
-        self.line_7.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line_7.setObjectName("line_7")
-        self.lineEdit_10 = QtWidgets.QLineEdit(self.scheda_dettaglio)
-        self.lineEdit_10.setGeometry(QtCore.QRect(1640, 10, 171, 25))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.lineEdit_10.setFont(font)
-        self.lineEdit_10.setText("")
-        self.lineEdit_10.setFrame(False)
-        self.lineEdit_10.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.lineEdit_10.setReadOnly(True)
-        self.lineEdit_10.setObjectName("lineEdit_10")
-        self.label_32 = QtWidgets.QLabel(self.scheda_dettaglio)
-        self.label_32.setGeometry(QtCore.QRect(1490, 15, 81, 21))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_32.setFont(font)
-        self.label_32.setObjectName("label_32")
-        self.label_33 = QtWidgets.QLabel(self.scheda_dettaglio)
-        self.label_33.setGeometry(QtCore.QRect(1490, 52, 81, 21))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_33.setFont(font)
-        self.label_33.setObjectName("label_33")
-        self.lineEdit_11 = QtWidgets.QLineEdit(self.scheda_dettaglio)
-        self.lineEdit_11.setGeometry(QtCore.QRect(1640, 50, 171, 25))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.lineEdit_11.setFont(font)
-        self.lineEdit_11.setText("")
-        self.lineEdit_11.setFrame(False)
-        self.lineEdit_11.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.lineEdit_11.setReadOnly(True)
-        self.lineEdit_11.setObjectName("lineEdit_11")
-        self.btn_deduci_descr = QtWidgets.QPushButton(self.scheda_dettaglio)
-        self.btn_deduci_descr.setGeometry(QtCore.QRect(1046, 470, 25, 25))
-        self.btn_deduci_descr.setText("")
-        icon2 = QtGui.QIcon()
-        icon2.addPixmap(QtGui.QPixmap("../Progetto/Risorse/Bottoni/reset.png"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
-        self.btn_deduci_descr.setIcon(icon2)
-        self.btn_deduci_descr.setObjectName("btn_deduci_descr")
-        self.line = QtWidgets.QFrame(self.scheda_dettaglio)
-        self.line.setGeometry(QtCore.QRect(1470, 470, 401, 20))
-        self.line.setFrameShape(QtWidgets.QFrame.HLine)
-        self.line.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line.setObjectName("line")
-        self.horizontalLayoutWidget_2 = QtWidgets.QWidget(self.scheda_dettaglio)
-        self.horizontalLayoutWidget_2.setGeometry(QtCore.QRect(430, 570, 371, 181))
-        self.horizontalLayoutWidget_2.setObjectName("horizontalLayoutWidget_2")
-        self.horizontalLayout_2 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_2)
-        self.horizontalLayout_2.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_2.setSpacing(5)
-        self.horizontalLayout_2.setObjectName("horizontalLayout_2")
-        self.frame_7 = QtWidgets.QFrame(self.horizontalLayoutWidget_2)
-        self.frame_7.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame_7.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame_7.setObjectName("frame_7")
-        self.label_foto_3d = QtWidgets.QLabel(self.frame_7)
-        self.label_foto_3d.setEnabled(False)
-        self.label_foto_3d.setGeometry(QtCore.QRect(1, 1, 177, 177))
-        self.label_foto_3d.setText("")
-        self.label_foto_3d.setPixmap(QtGui.QPixmap("../../../.designer/Risorse/regi_0.0.png"))
-        self.label_foto_3d.setScaledContents(True)
-        self.label_foto_3d.setWordWrap(False)
-        self.label_foto_3d.setObjectName("label_foto_3d")
-        self.horizontalLayout_2.addWidget(self.frame_7)
-        self.frame_6 = QtWidgets.QFrame(self.horizontalLayoutWidget_2)
-        self.frame_6.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame_6.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame_6.setObjectName("frame_6")
-        self.label_foto_2d = QtWidgets.QLabel(self.frame_6)
-        self.label_foto_2d.setGeometry(QtCore.QRect(0, 1, 177, 177))
-        self.label_foto_2d.setText("")
-        self.label_foto_2d.setPixmap(QtGui.QPixmap("../../../.designer/Risorse/regi_0.0.png"))
-        self.label_foto_2d.setScaledContents(True)
-        self.label_foto_2d.setObjectName("label_foto_2d")
-        self.horizontalLayout_2.addWidget(self.frame_6)
-        self.gridLayoutWidget = QtWidgets.QWidget(self.scheda_dettaglio)
-        self.gridLayoutWidget.setGeometry(QtCore.QRect(1470, 500, 391, 251))
-        self.gridLayoutWidget.setObjectName("gridLayoutWidget")
-        self.gridLayout = QtWidgets.QGridLayout(self.gridLayoutWidget)
-        self.gridLayout.setContentsMargins(0, 0, 0, 0)
-        self.gridLayout.setObjectName("gridLayout")
-        self.btn_salva_prev = QtWidgets.QPushButton(self.gridLayoutWidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.btn_salva_prev.sizePolicy().hasHeightForWidth())
-        self.btn_salva_prev.setSizePolicy(sizePolicy)
-        self.btn_salva_prev.setMinimumSize(QtCore.QSize(200, 40))
-        self.btn_salva_prev.setObjectName("btn_salva_prev")
-        self.gridLayout.addWidget(self.btn_salva_prev, 0, 0, 1, 1)
-        self.horizontalLayoutWidget_6 = QtWidgets.QWidget(self.scheda_dettaglio)
-        self.horizontalLayoutWidget_6.setGeometry(QtCore.QRect(10, 10, 391, 80))
-        self.horizontalLayoutWidget_6.setObjectName("horizontalLayoutWidget_6")
-        self.horizontalLayout_6 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_6)
-        self.horizontalLayout_6.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_6.setObjectName("horizontalLayout_6")
-        self.btn_nuovo_serr = QtWidgets.QPushButton(self.horizontalLayoutWidget_6)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.btn_nuovo_serr.sizePolicy().hasHeightForWidth())
-        self.btn_nuovo_serr.setSizePolicy(sizePolicy)
-        self.btn_nuovo_serr.setMinimumSize(QtCore.QSize(118, 50))
-        self.btn_nuovo_serr.setObjectName("btn_nuovo_serr")
-        self.horizontalLayout_6.addWidget(self.btn_nuovo_serr)
-        self.btn_dupl_serr = QtWidgets.QPushButton(self.horizontalLayoutWidget_6)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.btn_dupl_serr.sizePolicy().hasHeightForWidth())
-        self.btn_dupl_serr.setSizePolicy(sizePolicy)
-        self.btn_dupl_serr.setMinimumSize(QtCore.QSize(118, 50))
-        self.btn_dupl_serr.setObjectName("btn_dupl_serr")
-        self.horizontalLayout_6.addWidget(self.btn_dupl_serr)
-        self.btn_canc_serr = QtWidgets.QPushButton(self.horizontalLayoutWidget_6)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(7)
-        sizePolicy.setHeightForWidth(self.btn_canc_serr.sizePolicy().hasHeightForWidth())
-        self.btn_canc_serr.setSizePolicy(sizePolicy)
-        self.btn_canc_serr.setMinimumSize(QtCore.QSize(118, 50))
-        self.btn_canc_serr.setObjectName("btn_canc_serr")
-        self.horizontalLayout_6.addWidget(self.btn_canc_serr)
-        self.verticalLayoutWidget = QtWidgets.QWidget(self.scheda_dettaglio)
-        self.verticalLayoutWidget.setGeometry(QtCore.QRect(1469, 120, 391, 341))
-        self.verticalLayoutWidget.setObjectName("verticalLayoutWidget")
-        self.verticalLayout = QtWidgets.QVBoxLayout(self.verticalLayoutWidget)
-        self.verticalLayout.setContentsMargins(0, 0, 0, 0)
-        self.verticalLayout.setSpacing(0)
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.frame_22 = QtWidgets.QFrame(self.verticalLayoutWidget)
-        self.frame_22.setObjectName("frame_22")
-        self.radio_prezzo_listino = QtWidgets.QRadioButton(self.frame_22)
-        self.radio_prezzo_listino.setEnabled(False)
-        self.radio_prezzo_listino.setGeometry(QtCore.QRect(5, 15, 141, 17))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.radio_prezzo_listino.setFont(font)
-        self.radio_prezzo_listino.setObjectName("radio_prezzo_listino")
-        self.radio_prezzo_custom = QtWidgets.QRadioButton(self.frame_22)
-        self.radio_prezzo_custom.setGeometry(QtCore.QRect(5, 45, 171, 17))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.radio_prezzo_custom.setFont(font)
-        self.radio_prezzo_custom.setChecked(True)
-        self.radio_prezzo_custom.setObjectName("radio_prezzo_custom")
-        self.lineEdit_2 = QtWidgets.QLineEdit(self.frame_22)
-        self.lineEdit_2.setGeometry(QtCore.QRect(351, 40, 25, 25))
-        self.lineEdit_2.setFrame(False)
-        self.lineEdit_2.setAlignment(QtCore.Qt.AlignCenter)
-        self.lineEdit_2.setReadOnly(True)
-        self.lineEdit_2.setObjectName("lineEdit_2")
-        self.lineEdit = QtWidgets.QLineEdit(self.frame_22)
-        self.lineEdit.setEnabled(False)
-        self.lineEdit.setGeometry(QtCore.QRect(351, 10, 25, 25))
-        self.lineEdit.setText("")
-        self.lineEdit.setFrame(False)
-        self.lineEdit.setAlignment(QtCore.Qt.AlignCenter)
-        self.lineEdit.setReadOnly(True)
-        self.lineEdit.setObjectName("lineEdit")
-        self.edit_cod_prog_7 = QtWidgets.QLineEdit(self.frame_22)
-        self.edit_cod_prog_7.setEnabled(True)
-        self.edit_cod_prog_7.setGeometry(QtCore.QRect(181, 40, 151, 25))
-        self.edit_cod_prog_7.setText("")
-        self.edit_cod_prog_7.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.edit_cod_prog_7.setReadOnly(True)
-        self.edit_cod_prog_7.setPlaceholderText("")
-        self.edit_cod_prog_7.setObjectName("edit_cod_prog_7")
-        self.edit_cod_prog_2 = QtWidgets.QLineEdit(self.frame_22)
-        self.edit_cod_prog_2.setEnabled(False)
-        self.edit_cod_prog_2.setGeometry(QtCore.QRect(180, 10, 151, 25))
-        self.edit_cod_prog_2.setText("")
-        self.edit_cod_prog_2.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.edit_cod_prog_2.setReadOnly(True)
-        self.edit_cod_prog_2.setPlaceholderText("")
-        self.edit_cod_prog_2.setObjectName("edit_cod_prog_2")
-        self.label_13 = QtWidgets.QLabel(self.frame_22)
-        self.label_13.setEnabled(False)
-        self.label_13.setGeometry(QtCore.QRect(336, 15, 16, 16))
-        self.label_13.setObjectName("label_13")
-        self.label_34 = QtWidgets.QLabel(self.frame_22)
-        self.label_34.setGeometry(QtCore.QRect(336, 45, 16, 16))
-        self.label_34.setObjectName("label_34")
-        self.verticalLayout.addWidget(self.frame_22)
-        self.frame_21 = QtWidgets.QFrame(self.verticalLayoutWidget)
-        self.frame_21.setObjectName("frame_21")
-        self.label_38 = QtWidgets.QLabel(self.frame_21)
-        self.label_38.setGeometry(QtCore.QRect(335, 55, 16, 16))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_38.setFont(font)
-        self.label_38.setObjectName("label_38")
-        self.edit_cod_prog_4 = QtWidgets.QLineEdit(self.frame_21)
-        self.edit_cod_prog_4.setEnabled(False)
-        self.edit_cod_prog_4.setGeometry(QtCore.QRect(250, 50, 81, 25))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setWeight(50)
-        self.edit_cod_prog_4.setFont(font)
-        self.edit_cod_prog_4.setText("")
-        self.edit_cod_prog_4.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.edit_cod_prog_4.setReadOnly(True)
-        self.edit_cod_prog_4.setPlaceholderText("")
-        self.edit_cod_prog_4.setObjectName("edit_cod_prog_4")
-        self.radio_sconto = QtWidgets.QRadioButton(self.frame_21)
-        self.radio_sconto.setGeometry(QtCore.QRect(5, 15, 81, 17))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.radio_sconto.setFont(font)
-        self.radio_sconto.setChecked(True)
-        self.radio_sconto.setObjectName("radio_sconto")
-        self.radio_sconto_percent = QtWidgets.QRadioButton(self.frame_21)
-        self.radio_sconto_percent.setGeometry(QtCore.QRect(5, 55, 131, 17))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.radio_sconto_percent.setFont(font)
-        self.radio_sconto_percent.setObjectName("radio_sconto_percent")
-        self.edit_cod_prog_3 = QtWidgets.QLineEdit(self.frame_21)
-        self.edit_cod_prog_3.setEnabled(True)
-        self.edit_cod_prog_3.setGeometry(QtCore.QRect(250, 10, 81, 25))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setWeight(50)
-        self.edit_cod_prog_3.setFont(font)
-        self.edit_cod_prog_3.setText("")
-        self.edit_cod_prog_3.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.edit_cod_prog_3.setReadOnly(True)
-        self.edit_cod_prog_3.setPlaceholderText("")
-        self.edit_cod_prog_3.setObjectName("edit_cod_prog_3")
-        self.verticalLayout.addWidget(self.frame_21)
-        self.frame_19 = QtWidgets.QFrame(self.verticalLayoutWidget)
-        self.frame_19.setObjectName("frame_19")
-        self.label_39 = QtWidgets.QLabel(self.frame_19)
-        self.label_39.setGeometry(QtCore.QRect(335, 33, 20, 20))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_39.setFont(font)
-        self.label_39.setObjectName("label_39")
-        self.label_37 = QtWidgets.QLabel(self.frame_19)
-        self.label_37.setGeometry(QtCore.QRect(25, 30, 41, 16))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_37.setFont(font)
-        self.label_37.setObjectName("label_37")
-        self.edit_iva = QtWidgets.QLineEdit(self.frame_19)
-        self.edit_iva.setGeometry(QtCore.QRect(250, 30, 81, 25))
-        self.edit_iva.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.edit_iva.setReadOnly(True)
-        self.edit_iva.setPlaceholderText("")
-        self.edit_iva.setObjectName("edit_iva")
-        self.verticalLayout.addWidget(self.frame_19)
-        self.frame_20 = QtWidgets.QFrame(self.verticalLayoutWidget)
-        self.frame_20.setObjectName("frame_20")
-        self.edit_cod_prog_5 = QtWidgets.QLineEdit(self.frame_20)
-        self.edit_cod_prog_5.setGeometry(QtCore.QRect(190, 30, 141, 25))
-        self.edit_cod_prog_5.setText("")
-        self.edit_cod_prog_5.setFrame(False)
-        self.edit_cod_prog_5.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.edit_cod_prog_5.setReadOnly(True)
-        self.edit_cod_prog_5.setPlaceholderText("")
-        self.edit_cod_prog_5.setObjectName("edit_cod_prog_5")
-        self.label_35 = QtWidgets.QLabel(self.frame_20)
-        self.label_35.setGeometry(QtCore.QRect(25, 32, 111, 21))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_35.setFont(font)
-        self.label_35.setObjectName("label_35")
-        self.verticalLayout.addWidget(self.frame_20)
-        self.verticalLayoutWidget_2 = QtWidgets.QWidget(self.scheda_dettaglio)
-        self.verticalLayoutWidget_2.setGeometry(QtCore.QRect(1110, 10, 321, 741))
-        self.verticalLayoutWidget_2.setObjectName("verticalLayoutWidget_2")
-        self.verticalLayout_2 = QtWidgets.QVBoxLayout(self.verticalLayoutWidget_2)
-        self.verticalLayout_2.setContentsMargins(0, 0, 0, 0)
-        self.verticalLayout_2.setSpacing(0)
-        self.verticalLayout_2.setObjectName("verticalLayout_2")
-        self.frame_24 = QtWidgets.QFrame(self.verticalLayoutWidget_2)
-        self.frame_24.setObjectName("frame_24")
-        self.label_12 = QtWidgets.QLabel(self.frame_24)
-        self.label_12.setGeometry(QtCore.QRect(10, 10, 71, 21))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_12.setFont(font)
-        self.label_12.setObjectName("label_12")
-        self.spin_num_pz = QtWidgets.QSpinBox(self.frame_24)
-        self.spin_num_pz.setGeometry(QtCore.QRect(140, 10, 151, 25))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setWeight(50)
-        self.spin_num_pz.setFont(font)
-        self.spin_num_pz.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.spin_num_pz.setReadOnly(False)
-        self.spin_num_pz.setPrefix("")
-        self.spin_num_pz.setMinimum(1)
-        self.spin_num_pz.setMaximum(10)
-        self.spin_num_pz.setSingleStep(1)
-        self.spin_num_pz.setObjectName("spin_num_pz")
-        self.verticalLayout_2.addWidget(self.frame_24)
-        self.frame_23 = QtWidgets.QFrame(self.verticalLayoutWidget_2)
-        self.frame_23.setObjectName("frame_23")
-        self.label_9 = QtWidgets.QLabel(self.frame_23)
-        self.label_9.setGeometry(QtCore.QRect(10, 15, 81, 16))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_9.setFont(font)
-        self.label_9.setObjectName("label_9")
-        self.label_10 = QtWidgets.QLabel(self.frame_23)
-        self.label_10.setGeometry(QtCore.QRect(10, 55, 61, 16))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_10.setFont(font)
-        self.label_10.setObjectName("label_10")
-        self.label_11 = QtWidgets.QLabel(self.frame_23)
-        self.label_11.setEnabled(False)
-        self.label_11.setGeometry(QtCore.QRect(10, 92, 61, 20))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_11.setFont(font)
-        self.label_11.setObjectName("label_11")
-        self.spin_lunghezza = QtWidgets.QSpinBox(self.frame_23)
-        self.spin_lunghezza.setGeometry(QtCore.QRect(140, 10, 151, 25))
-        self.spin_lunghezza.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.spin_lunghezza.setPrefix("")
-        self.spin_lunghezza.setMaximum(5000)
-        self.spin_lunghezza.setSingleStep(10)
-        self.spin_lunghezza.setObjectName("spin_lunghezza")
-        self.spin_altezza = QtWidgets.QSpinBox(self.frame_23)
-        self.spin_altezza.setGeometry(QtCore.QRect(140, 50, 151, 25))
-        font = QtGui.QFont()
-        font.setItalic(False)
-        font.setUnderline(False)
-        font.setStrikeOut(False)
-        font.setKerning(True)
-        self.spin_altezza.setFont(font)
-        self.spin_altezza.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.spin_altezza.setPrefix("")
-        self.spin_altezza.setMaximum(5000)
-        self.spin_altezza.setSingleStep(10)
-        self.spin_altezza.setObjectName("spin_altezza")
-        self.spin_spessore = QtWidgets.QSpinBox(self.frame_23)
-        self.spin_spessore.setEnabled(False)
-        self.spin_spessore.setGeometry(QtCore.QRect(140, 90, 151, 25))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setItalic(False)
-        font.setUnderline(False)
-        font.setWeight(50)
-        font.setStrikeOut(False)
-        font.setKerning(True)
-        self.spin_spessore.setFont(font)
-        self.spin_spessore.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.spin_spessore.setPrefix("")
-        self.spin_spessore.setMaximum(10000)
-        self.spin_spessore.setSingleStep(10)
-        self.spin_spessore.setObjectName("spin_spessore")
-        self.verticalLayout_2.addWidget(self.frame_23)
-        self.frame_4 = QtWidgets.QFrame(self.verticalLayoutWidget_2)
-        self.frame_4.setObjectName("frame_4")
-        self.label_28 = QtWidgets.QLabel(self.frame_4)
-        self.label_28.setGeometry(QtCore.QRect(10, 15, 81, 21))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_28.setFont(font)
-        self.label_28.setObjectName("label_28")
-        self.label_29 = QtWidgets.QLabel(self.frame_4)
-        self.label_29.setGeometry(QtCore.QRect(10, 55, 91, 21))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_29.setFont(font)
-        self.label_29.setObjectName("label_29")
-        self.label_30 = QtWidgets.QLabel(self.frame_4)
-        self.label_30.setEnabled(False)
-        self.label_30.setGeometry(QtCore.QRect(10, 92, 71, 21))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_30.setFont(font)
-        self.label_30.setObjectName("label_30")
-        self.combo_materiale = QtWidgets.QComboBox(self.frame_4)
-        self.combo_materiale.setGeometry(QtCore.QRect(120, 10, 171, 25))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setWeight(50)
-        self.combo_materiale.setFont(font)
-        self.combo_materiale.setFrame(True)
-        self.combo_materiale.setObjectName("combo_materiale")
-        self.combo_materiale.addItem("")
-        self.combo_materiale.addItem("")
-        self.combo_vernice = QtWidgets.QComboBox(self.frame_4)
-        self.combo_vernice.setGeometry(QtCore.QRect(120, 50, 171, 25))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setWeight(50)
-        self.combo_vernice.setFont(font)
-        self.combo_vernice.setObjectName("combo_vernice")
-        self.combo_vernice.addItem("")
-        self.combo_variante3 = QtWidgets.QComboBox(self.frame_4)
-        self.combo_variante3.setEnabled(False)
-        self.combo_variante3.setGeometry(QtCore.QRect(120, 90, 171, 25))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setWeight(50)
-        self.combo_variante3.setFont(font)
-        self.combo_variante3.setObjectName("combo_variante3")
-        self.verticalLayout_2.addWidget(self.frame_4)
-        self.frame_5 = QtWidgets.QFrame(self.verticalLayoutWidget_2)
-        self.frame_5.setObjectName("frame_5")
-        self.label_31 = QtWidgets.QLabel(self.frame_5)
-        self.label_31.setGeometry(QtCore.QRect(10, 0, 111, 21))
-        font = QtGui.QFont()
-        font.setPointSize(8)
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_31.setFont(font)
-        self.label_31.setObjectName("label_31")
-        self.edit_note_varianti = QtWidgets.QTextEdit(self.frame_5)
-        self.edit_note_varianti.setGeometry(QtCore.QRect(10, 30, 281, 151))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        font.setBold(False)
-        font.setWeight(50)
-        self.edit_note_varianti.setFont(font)
-        self.edit_note_varianti.setObjectName("edit_note_varianti")
-        self.verticalLayout_2.addWidget(self.frame_5)
-        self.tree_progetti = QtWidgets.QTreeView(self.scheda_dettaglio)
-        self.tree_progetti.setEnabled(True)
-        self.tree_progetti.setAnimated(True)
-        self.tree_progetti.setGeometry(QtCore.QRect(430, 50, 641, 380))
-        self.tree_progetti.setObjectName("tree_progetti")
-        icon3 = QtGui.QIcon()
-        icon3.addPixmap(QtGui.QPixmap("../Progetto/Risorse/Bottoni/dettaglio.png"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
-        self.tab_dettaglio_totale_stampa.addTab(self.scheda_dettaglio, icon3, "")
-        self.scheda_totale = QtWidgets.QWidget()
-        self.scheda_totale.setObjectName("scheda_totale")
-        self.label_36 = QtWidgets.QLabel(self.scheda_totale)
-        self.label_36.setGeometry(QtCore.QRect(10, 15, 121, 21))
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setWeight(50)
-        self.label_36.setFont(font)
-        self.label_36.setObjectName("label_36")
-        self.tableWidget = QtWidgets.QTableWidget(self.scheda_totale)
-        self.tableWidget.setGeometry(QtCore.QRect(10, 50, 400, 701))
-        self.tableWidget.setAlternatingRowColors(True)
-        self.tableWidget.setObjectName("tableWidget")
-        self.tableWidget.setColumnCount(3)
-        self.tableWidget.setRowCount(0)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setHorizontalHeaderItem(0, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setHorizontalHeaderItem(1, item)
-        item = QtWidgets.QTableWidgetItem()
-        self.tableWidget.setHorizontalHeaderItem(2, item)
-        self.tableWidget.horizontalHeader().setDefaultSectionSize(141)
-        self.tableWidget.verticalHeader().setVisible(False)
-        self.line_8 = QtWidgets.QFrame(self.scheda_totale)
-        self.line_8.setGeometry(QtCore.QRect(730, 0, 20, 751))
-        self.line_8.setFrameShape(QtWidgets.QFrame.VLine)
-        self.line_8.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line_8.setObjectName("line_8")
-        icon4 = QtGui.QIcon()
-        icon4.addPixmap(QtGui.QPixmap("../Progetto/Risorse/Bottoni/calc.png"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
-        self.tab_dettaglio_totale_stampa.addTab(self.scheda_totale, icon4, "")
-        self.frame = QtWidgets.QFrame(self.pag_2)
-        self.frame.setGeometry(QtCore.QRect(10, 10, 1881, 130))
-        self.frame.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame.setObjectName("frame")
-        self.label = QtWidgets.QLabel(self.frame)
-        self.label.setGeometry(QtCore.QRect(10, 10, 101, 16))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label.setFont(font)
-        self.label.setObjectName("label")
-        self.edit_so = QtWidgets.QLineEdit(self.frame)
-        self.edit_so.setGeometry(QtCore.QRect(112, 30, 101, 25))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.edit_so.setFont(font)
-        self.edit_so.setText("")
-        self.edit_so.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.edit_so.setObjectName("edit_so")
-        self.label_5 = QtWidgets.QLabel(self.frame)
-        self.label_5.setGeometry(QtCore.QRect(270, 10, 41, 16))
-        self.label_5.setObjectName("label_5")
-        self.edit_data_prev = QtWidgets.QDateEdit(self.frame)
-        self.edit_data_prev.setGeometry(QtCore.QRect(270, 30, 161, 25))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        self.edit_data_prev.setFont(font)
-        self.edit_data_prev.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.edit_data_prev.setMaximumDate(QtCore.QDate(2049, 12, 31))
-        self.edit_data_prev.setMinimumDate(QtCore.QDate(2019, 1, 1))
-        self.edit_data_prev.setDate(QtCore.QDate(2019, 12, 4))
-        self.edit_data_prev.setObjectName("edit_data_prev")
-        self.line_3 = QtWidgets.QFrame(self.frame)
-        self.line_3.setGeometry(QtCore.QRect(940, 10, 20, 111))
-        self.line_3.setFrameShape(QtWidgets.QFrame.VLine)
-        self.line_3.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line_3.setObjectName("line_3")
-        self.edit_testo_in = QtWidgets.QLineEdit(self.frame)
-        self.edit_testo_in.setGeometry(QtCore.QRect(489, 30, 381, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.edit_testo_in.setFont(font)
-        self.edit_testo_in.setReadOnly(True)
-        self.edit_testo_in.setObjectName("edit_testo_in")
-        self.btn_scegli_testo_in = QtWidgets.QToolButton(self.frame)
-        self.btn_scegli_testo_in.setGeometry(QtCore.QRect(880, 30, 25, 25))
-        self.btn_scegli_testo_in.setCheckable(False)
-        self.btn_scegli_testo_in.setChecked(False)
-        self.btn_scegli_testo_in.setArrowType(QtCore.Qt.NoArrow)
-        self.btn_scegli_testo_in.setObjectName("btn_scegli_testo_in")
-        self.label_15 = QtWidgets.QLabel(self.frame)
-        self.label_15.setGeometry(QtCore.QRect(490, 10, 121, 16))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_15.setFont(font)
-        self.label_15.setObjectName("label_15")
-        self.edit_testo_fin = QtWidgets.QLineEdit(self.frame)
-        self.edit_testo_fin.setGeometry(QtCore.QRect(489, 90, 381, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.edit_testo_fin.setFont(font)
-        self.edit_testo_fin.setReadOnly(True)
-        self.edit_testo_fin.setObjectName("edit_testo_fin")
-        self.btn_scegli_testo_fin = QtWidgets.QToolButton(self.frame)
-        self.btn_scegli_testo_fin.setGeometry(QtCore.QRect(880, 90, 25, 25))
-        self.btn_scegli_testo_fin.setObjectName("btn_scegli_testo_fin")
-        self.label_16 = QtWidgets.QLabel(self.frame)
-        self.label_16.setGeometry(QtCore.QRect(490, 70, 111, 16))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.label_16.setFont(font)
-        self.label_16.setObjectName("label_16")
-        self.edit_nick_cliente = QtWidgets.QLineEdit(self.frame)
-        self.edit_nick_cliente.setGeometry(QtCore.QRect(270, 90, 161, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.edit_nick_cliente.setFont(font)
-        self.edit_nick_cliente.setText("")
-        self.edit_nick_cliente.setObjectName("edit_nick_cliente")
-        self.label_20 = QtWidgets.QLabel(self.frame)
-        self.label_20.setGeometry(QtCore.QRect(270, 70, 151, 16))
-        self.label_20.setObjectName("label_20")
-        self.btn_deduci_testo_in = QtWidgets.QPushButton(self.frame)
-        self.btn_deduci_testo_in.setGeometry(QtCore.QRect(910, 30, 25, 25))
-        self.btn_deduci_testo_in.setText("")
-        self.btn_deduci_testo_in.setIcon(icon2)
-        self.btn_deduci_testo_in.setObjectName("btn_deduci_testo_in")
-        self.btn_deduci_data = QtWidgets.QPushButton(self.frame)
-        self.btn_deduci_data.setGeometry(QtCore.QRect(440, 30, 25, 25))
-        self.btn_deduci_data.setText("")
-        self.btn_deduci_data.setIcon(icon2)
-        self.btn_deduci_data.setObjectName("btn_deduci_data")
-        self.btn_deduci_nick_cliente = QtWidgets.QPushButton(self.frame)
-        self.btn_deduci_nick_cliente.setGeometry(QtCore.QRect(440, 90, 25, 25))
-        self.btn_deduci_nick_cliente.setText("")
-        self.btn_deduci_nick_cliente.setIcon(icon2)
-        self.btn_deduci_nick_cliente.setObjectName("btn_deduci_nick_cliente")
-        self.btn_deduci_testo_fin = QtWidgets.QPushButton(self.frame)
-        self.btn_deduci_testo_fin.setGeometry(QtCore.QRect(910, 90, 25, 25))
-        self.btn_deduci_testo_fin.setText("")
-        self.btn_deduci_testo_fin.setIcon(icon2)
-        self.btn_deduci_testo_fin.setObjectName("btn_deduci_testo_fin")
-        self.btn_deduci_po = QtWidgets.QPushButton(self.frame)
-        self.btn_deduci_po.setEnabled(False)
-        self.btn_deduci_po.setGeometry(QtCore.QRect(220, 90, 25, 25))
-        self.btn_deduci_po.setText("")
-        self.btn_deduci_po.setIcon(icon2)
-        self.btn_deduci_po.setObjectName("btn_deduci_po")
-        self.edit_po = QtWidgets.QLineEdit(self.frame)
-        self.edit_po.setEnabled(False)
-        self.edit_po.setGeometry(QtCore.QRect(112, 90, 101, 25))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.edit_po.setFont(font)
-        self.edit_po.setText("")
-        self.edit_po.setAlignment(QtCore.Qt.AlignRight|QtCore.Qt.AlignTrailing|QtCore.Qt.AlignVCenter)
-        self.edit_po.setObjectName("edit_po")
-        self.btn_deduci_so = QtWidgets.QPushButton(self.frame)
-        self.btn_deduci_so.setGeometry(QtCore.QRect(220, 30, 25, 25))
-        self.btn_deduci_so.setText("")
-        self.btn_deduci_so.setIcon(icon2)
-        self.btn_deduci_so.setObjectName("btn_deduci_so")
-        self.frame_3 = QtWidgets.QFrame(self.frame)
-        self.frame_3.setGeometry(QtCore.QRect(10, 30, 101, 91))
-        self.frame_3.setObjectName("frame_3")
-        self.radio_so = QtWidgets.QRadioButton(self.frame_3)
-        self.radio_so.setGeometry(QtCore.QRect(0, 4, 101, 17))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.radio_so.setFont(font)
-        self.radio_so.setChecked(True)
-        self.radio_so.setObjectName("radio_so")
-        self.radio_po = QtWidgets.QRadioButton(self.frame_3)
-        self.radio_po.setEnabled(False)
-        self.radio_po.setGeometry(QtCore.QRect(0, 56, 101, 31))
-        font = QtGui.QFont()
-        font.setPointSize(10)
-        font.setBold(True)
-        font.setWeight(75)
-        self.radio_po.setFont(font)
-        self.radio_po.setObjectName("radio_po")
-        self.horizontalLayoutWidget_3 = QtWidgets.QWidget(self.frame)
-        self.horizontalLayoutWidget_3.setGeometry(QtCore.QRect(959, 10, 911, 51))
-        self.horizontalLayoutWidget_3.setObjectName("horizontalLayoutWidget_3")
-        self.horizontalLayout_3 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_3)
-        self.horizontalLayout_3.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_3.setObjectName("horizontalLayout_3")
-        self.frame_10 = QtWidgets.QFrame(self.horizontalLayoutWidget_3)
-        self.frame_10.setObjectName("frame_10")
-        self.label_6 = QtWidgets.QLabel(self.frame_10)
-        self.label_6.setGeometry(QtCore.QRect(0, 0, 131, 16))
-        self.label_6.setObjectName("label_6")
-        self.line_prev_nome = QtWidgets.QLineEdit(self.frame_10)
-        self.line_prev_nome.setGeometry(QtCore.QRect(0, 20, 291, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.line_prev_nome.setFont(font)
-        self.line_prev_nome.setText("")
-        self.line_prev_nome.setFrame(False)
-        self.line_prev_nome.setReadOnly(True)
-        self.line_prev_nome.setObjectName("line_prev_nome")
-        self.horizontalLayout_3.addWidget(self.frame_10)
-        self.frame_9 = QtWidgets.QFrame(self.horizontalLayoutWidget_3)
-        self.frame_9.setObjectName("frame_9")
-        self.label_7 = QtWidgets.QLabel(self.frame_9)
-        self.label_7.setGeometry(QtCore.QRect(0, 0, 61, 16))
-        self.label_7.setObjectName("label_7")
-        self.line_prev_via = QtWidgets.QLineEdit(self.frame_9)
-        self.line_prev_via.setGeometry(QtCore.QRect(0, 20, 291, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.line_prev_via.setFont(font)
-        self.line_prev_via.setText("")
-        self.line_prev_via.setFrame(False)
-        self.line_prev_via.setReadOnly(True)
-        self.line_prev_via.setObjectName("line_prev_via")
-        self.horizontalLayout_3.addWidget(self.frame_9)
-        self.frame_8 = QtWidgets.QFrame(self.horizontalLayoutWidget_3)
-        self.frame_8.setObjectName("frame_8")
-        self.label_8 = QtWidgets.QLabel(self.frame_8)
-        self.label_8.setGeometry(QtCore.QRect(0, 0, 111, 16))
-        self.label_8.setObjectName("label_8")
-        self.line_prev_cantiere = QtWidgets.QLineEdit(self.frame_8)
-        self.line_prev_cantiere.setGeometry(QtCore.QRect(0, 20, 291, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.line_prev_cantiere.setFont(font)
-        self.line_prev_cantiere.setText("")
-        self.line_prev_cantiere.setFrame(False)
-        self.line_prev_cantiere.setReadOnly(True)
-        self.line_prev_cantiere.setObjectName("line_prev_cantiere")
-        self.horizontalLayout_3.addWidget(self.frame_8)
-        self.horizontalLayoutWidget_4 = QtWidgets.QWidget(self.frame)
-        self.horizontalLayoutWidget_4.setGeometry(QtCore.QRect(959, 70, 911, 51))
-        self.horizontalLayoutWidget_4.setObjectName("horizontalLayoutWidget_4")
-        self.horizontalLayout_4 = QtWidgets.QHBoxLayout(self.horizontalLayoutWidget_4)
-        self.horizontalLayout_4.setContentsMargins(0, 0, 0, 0)
-        self.horizontalLayout_4.setObjectName("horizontalLayout_4")
-        self.frame_14 = QtWidgets.QFrame(self.horizontalLayoutWidget_4)
-        self.frame_14.setObjectName("frame_14")
-        self.line_prev_comune = QtWidgets.QLineEdit(self.frame_14)
-        self.line_prev_comune.setGeometry(QtCore.QRect(0, 20, 211, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.line_prev_comune.setFont(font)
-        self.line_prev_comune.setText("")
-        self.line_prev_comune.setFrame(False)
-        self.line_prev_comune.setReadOnly(True)
-        self.line_prev_comune.setObjectName("line_prev_comune")
-        self.label_14 = QtWidgets.QLabel(self.frame_14)
-        self.label_14.setGeometry(QtCore.QRect(0, 0, 131, 16))
-        self.label_14.setObjectName("label_14")
-        self.horizontalLayout_4.addWidget(self.frame_14)
-        self.frame_13 = QtWidgets.QFrame(self.horizontalLayoutWidget_4)
-        self.frame_13.setObjectName("frame_13")
-        self.line_prev_tel1 = QtWidgets.QLineEdit(self.frame_13)
-        self.line_prev_tel1.setGeometry(QtCore.QRect(0, 20, 211, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.line_prev_tel1.setFont(font)
-        self.line_prev_tel1.setText("")
-        self.line_prev_tel1.setFrame(False)
-        self.line_prev_tel1.setReadOnly(True)
-        self.line_prev_tel1.setObjectName("line_prev_tel1")
-        self.label_17 = QtWidgets.QLabel(self.frame_13)
-        self.label_17.setGeometry(QtCore.QRect(0, 0, 111, 16))
-        self.label_17.setObjectName("label_17")
-        self.horizontalLayout_4.addWidget(self.frame_13)
-        self.frame_12 = QtWidgets.QFrame(self.horizontalLayoutWidget_4)
-        self.frame_12.setObjectName("frame_12")
-        self.line_prev_tel2 = QtWidgets.QLineEdit(self.frame_12)
-        self.line_prev_tel2.setGeometry(QtCore.QRect(0, 20, 211, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.line_prev_tel2.setFont(font)
-        self.line_prev_tel2.setText("")
-        self.line_prev_tel2.setFrame(False)
-        self.line_prev_tel2.setReadOnly(True)
-        self.line_prev_tel2.setObjectName("line_prev_tel2")
-        self.label_18 = QtWidgets.QLabel(self.frame_12)
-        self.label_18.setGeometry(QtCore.QRect(0, 0, 111, 16))
-        self.label_18.setObjectName("label_18")
-        self.horizontalLayout_4.addWidget(self.frame_12)
-        self.frame_11 = QtWidgets.QFrame(self.horizontalLayoutWidget_4)
-        self.frame_11.setObjectName("frame_11")
-        self.line_prev_email = QtWidgets.QLineEdit(self.frame_11)
-        self.line_prev_email.setGeometry(QtCore.QRect(0, 20, 211, 25))
-        font = QtGui.QFont()
-        font.setPointSize(9)
-        self.line_prev_email.setFont(font)
-        self.line_prev_email.setText("")
-        self.line_prev_email.setFrame(False)
-        self.line_prev_email.setReadOnly(True)
-        self.line_prev_email.setObjectName("line_prev_email")
-        self.label_19 = QtWidgets.QLabel(self.frame_11)
-        self.label_19.setGeometry(QtCore.QRect(0, 0, 111, 16))
-        self.label_19.setObjectName("label_19")
-        self.horizontalLayout_4.addWidget(self.frame_11)
-        self.stackedWidget.addWidget(self.pag_2)
-        self.pag_3 = QtWidgets.QWidget()
-        self.pag_3.setObjectName("pag_3")
-        self.stackedWidget.addWidget(self.pag_3)
-        reGI.setCentralWidget(self.centralWidget)
-        self.menubar = QtWidgets.QMenuBar(reGI)
-        self.menubar.setEnabled(True)
-        self.menubar.setGeometry(QtCore.QRect(0, 0, 1920, 20))
-        self.menubar.setObjectName("menubar")
-        self.menuNuovo = QtWidgets.QMenu(self.menubar)
-        self.menuNuovo.setObjectName("menuNuovo")
-        self.menuNuovo_2 = QtWidgets.QMenu(self.menuNuovo)
-        self.menuNuovo_2.setObjectName("menuNuovo_2")
-        self.menuInserisci = QtWidgets.QMenu(self.menubar)
-        self.menuInserisci.setObjectName("menuInserisci")
-        self.menuVista = QtWidgets.QMenu(self.menubar)
-        self.menuVista.setObjectName("menuVista")
-        self.menuSeleziona_stile = QtWidgets.QMenu(self.menuVista)
-        self.menuSeleziona_stile.setObjectName("menuSeleziona_stile")
-        self.menuPalette = QtWidgets.QMenu(self.menuVista)
-        self.menuPalette.setObjectName("menuPalette")
-        self.menuAnteprima = QtWidgets.QMenu(self.menubar)
-        self.menuAnteprima.setObjectName("menuAnteprima")
-        self.menuInfo = QtWidgets.QMenu(self.menubar)
-        self.menuInfo.setObjectName("menuInfo")
-        reGI.setMenuBar(self.menubar)
-        self.statusbar = QtWidgets.QStatusBar(reGI)
-        self.statusbar.setEnabled(True)
-        self.statusbar.setObjectName("statusbar")
-        reGI.setStatusBar(self.statusbar)
-        self.actionVersione = QtWidgets.QAction(reGI)
-        self.actionVersione.setObjectName("actionVersione")
-        self.actionVerifica_aggiornamenti = QtWidgets.QAction(reGI)
-        self.actionVerifica_aggiornamenti.setObjectName("actionVerifica_aggiornamenti")
-        self.actionApri = QtWidgets.QAction(reGI)
-        self.actionApri.setObjectName("actionApri")
-        self.actionRecente_1 = QtWidgets.QAction(reGI)
-        self.actionRecente_1.setObjectName("actionRecente_1")
-        self.actionRecente_2 = QtWidgets.QAction(reGI)
-        self.actionRecente_2.setObjectName("actionRecente_2")
-        self.actionDa_zero = QtWidgets.QAction(reGI)
-        self.actionDa_zero.setObjectName("actionDa_zero")
-        self.actionDa_precompilato = QtWidgets.QAction(reGI)
-        self.actionDa_precompilato.setEnabled(False)
-        self.actionDa_precompilato.setObjectName("actionDa_precompilato")
-        self.actionInfisso = QtWidgets.QAction(reGI)
-        self.actionInfisso.setEnabled(False)
-        self.actionInfisso.setObjectName("actionInfisso")
-        self.actionVisualizza_anteprima_di_stampa = QtWidgets.QAction(reGI)
-        self.actionVisualizza_anteprima_di_stampa.setEnabled(False)
-        self.actionVisualizza_anteprima_di_stampa.setObjectName("actionVisualizza_anteprima_di_stampa")
-        self.actionDefault = QtWidgets.QAction(reGI)
-        self.actionDefault.setObjectName("actionDefault")
-        self.actionFusion = QtWidgets.QAction(reGI)
-        self.actionFusion.setObjectName("actionFusion")
-        self.actionWindows = QtWidgets.QAction(reGI)
-        self.actionWindows.setObjectName("actionWindows")
-        self.actionWindows_Vista = QtWidgets.QAction(reGI)
-        self.actionWindows_Vista.setObjectName("actionWindows_Vista")
-        self.actionSchermata_iniziale = QtWidgets.QAction(reGI)
-        self.actionSchermata_iniziale.setCheckable(False)
-        self.actionSchermata_iniziale.setObjectName("actionSchermata_iniziale")
-        self.actionAnagrafica_clienti = QtWidgets.QAction(reGI)
-        self.actionAnagrafica_clienti.setObjectName("actionAnagrafica_clienti")
-        self.actionLista_progetti = QtWidgets.QAction(reGI)
-        self.actionLista_progetti.setObjectName("actionLista_progetti")
-        self.actionCliente = QtWidgets.QAction(reGI)
-        self.actionCliente.setObjectName("actionCliente")
-        self.actionProgetto = QtWidgets.QAction(reGI)
-        self.actionProgetto.setObjectName("actionProgetto")
-        self.actionServe_aiuto = QtWidgets.QAction(reGI)
-        self.actionServe_aiuto.setObjectName("actionServe_aiuto")
-        self.actionProgetto_2 = QtWidgets.QAction(reGI)
-        self.actionProgetto_2.setObjectName("actionProgetto_2")
-        self.actionCliente_2 = QtWidgets.QAction(reGI)
-        self.actionCliente_2.setObjectName("actionCliente_2")
-        self.actionEsci = QtWidgets.QAction(reGI)
-        self.actionEsci.setObjectName("actionEsci")
-        self.actionPalette_1 = QtWidgets.QAction(reGI)
-        self.actionPalette_1.setObjectName("actionPalette_1")
-        self.actionPalette_2 = QtWidgets.QAction(reGI)
-        self.actionPalette_2.setObjectName("actionPalette_2")
-        self.menuNuovo_2.addAction(self.actionDa_zero)
-        self.menuNuovo_2.addAction(self.actionDa_precompilato)
-        self.menuNuovo_2.addAction(self.actionCliente)
-        self.menuNuovo_2.addAction(self.actionProgetto)
-        self.menuNuovo.addAction(self.actionSchermata_iniziale)
-        self.menuNuovo.addAction(self.actionAnagrafica_clienti)
-        self.menuNuovo.addAction(self.actionLista_progetti)
-        self.menuNuovo.addSeparator()
-        self.menuNuovo.addAction(self.menuNuovo_2.menuAction())
-        self.menuNuovo.addAction(self.actionApri)
-        self.menuNuovo.addSeparator()
-        self.menuNuovo.addAction(self.actionEsci)
-        self.menuInserisci.addAction(self.actionCliente_2)
-        self.menuInserisci.addAction(self.actionInfisso)
-        self.menuSeleziona_stile.addAction(self.actionDefault)
-        self.menuSeleziona_stile.addAction(self.actionFusion)
-        self.menuSeleziona_stile.addAction(self.actionWindows)
-        self.menuSeleziona_stile.addAction(self.actionWindows_Vista)
-        self.menuPalette.addAction(self.actionPalette_1)
-        self.menuPalette.addAction(self.actionPalette_2)
-        self.menuVista.addAction(self.menuSeleziona_stile.menuAction())
-        self.menuVista.addAction(self.menuPalette.menuAction())
-        self.menuAnteprima.addAction(self.actionVisualizza_anteprima_di_stampa)
-        self.menuInfo.addAction(self.actionServe_aiuto)
-        self.menuInfo.addSeparator()
-        self.menuInfo.addAction(self.actionVersione)
-        self.menuInfo.addAction(self.actionVerifica_aggiornamenti)
-        self.menubar.addAction(self.menuNuovo.menuAction())
-        self.menubar.addAction(self.menuInserisci.menuAction())
-        self.menubar.addAction(self.menuVista.menuAction())
-        self.menubar.addAction(self.menuAnteprima.menuAction())
-        self.menubar.addAction(self.menuInfo.menuAction())
-
-        self.retranslateUi(reGI)
-        self.stackedWidget.setCurrentIndex(2)
-        self.tab_dettaglio_totale_stampa.setCurrentIndex(0)
-        QtCore.QMetaObject.connectSlotsByName(reGI)
-
-        # <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> #
-
-        self.stackedWidget.setCurrentIndex(0)
-        self.current_cliente = {"id": 0, "cognome": "", "nome": "", "via": "", "comune": "", "tel1": "", "tel2": "", "email": "", "cantiere": ""}
-        self.current_prev = {"id": 0, "cliente": 0, "nick_cliente": "", "anno": 0, "so": 0, "po": 0, "data": "", "testo_in": "", "testo_out": ""}
-        self.lista_id_inf = []
-        self.current_infisso = {"id": 0, "prev": 0, "cod_prog": 0, "posizione": "", "descrizione": "", "note": "", "num_pz": 0, "lunghezza": 0,
-                                "altezza": 0, "spessore": 0, "materiale": "", "vernice": "", "note_varianti": "", "prezzo_netto": 0,
-                                "sconto": 0.0}
-        
-        self.btn_nuovo_prev.clicked.connect(self.view_db_clienti)
-        self.btn_view_anagrafica.clicked.connect(self.view_anagrafica)
-        self.btn_sc_agg_cliente.clicked.connect(self.sc_agg_cliente)
-        self.btn_sc_edit_cliente.clicked.connect(self.sc_edit_cliente)
-
-        self.btn_scegli_cliente.clicked.connect(self.crea_prev)
-        self.btn_agg_cliente.clicked.connect(self.agg_cliente)
-        self.btn_edit_cliente.clicked.connect(self.edit_cliente)
-        self.btn_canc_cliente.clicked.connect(self.canc_cliente)
-
-        self.btn_deduci_so.clicked.connect(self.deduci_SO)
-        self.btn_deduci_po.clicked.connect(self.deduci_PO)
-        self.btn_deduci_data.clicked.connect(self.deduci_data)
-        self.btn_deduci_nick_cliente.clicked.connect(self.deduci_nick_cliente)
-        self.btn_deduci_testo_in.clicked.connect(self.deduci_testo_in)
-        self.btn_scegli_testo_in.clicked.connect(self.scegli_testo_in)
-        self.btn_deduci_testo_fin.clicked.connect(self.deduci_testo_fin)
-        self.btn_scegli_testo_fin.clicked.connect(self.scegli_testo_fin)
-        self.btn_scegli_prog.clicked.connect(self.scegli_progetto)
-        self.btn_deduci_descr.clicked.connect(self.deduci_descr)
-        self.listwidget_inf.itemClicked.connect(self.load_infisso)
-
-        self.modello_tree_progetti = QtGui.QStandardItemModel()
-        self.tree_progetti.setModel(self.modello_tree_progetti)
-        self.root_node = self.modello_tree_progetti.invisibleRootItem()
-
-    def view_db_clienti(self):
-        self.stackedWidget.setCurrentIndex(1)
-        lista_clienti = Cliente.select().dicts()
-        for num_riga, cliente in enumerate(lista_clienti):
-            self.tab_clienti.insertRow(num_riga)
-            for num_colonna, dato in enumerate(cliente.values()):
-                if type(dato) is int:
-                    dato = str(dato)
-                    dato = dato.zfill(4)
-                if dato == None or dato == "":
-                    dato = "N/A"
-                self.tab_clienti.setItem(num_riga, num_colonna, QtWidgets.QTableWidgetItem(dato))
-    
-    def view_anagrafica(self):
-        self.view_db_clienti()
-        self.label_2.enabled = False
-        self.btn_scegli_cliente.enabled = False
-    
-    def sc_agg_cliente(self):
-        self.view_db_clienti()
-        self.label_2.enabled = False
-        self.btn_scegli_cliente.enabled = False
-        self.btn_edit_cliente.enabled = False
-        self.btn_canc_cliente.enabled = False
-    
-    def sc_edit_cliente(self):
-        self.view_db_clienti()
-        self.label_2.enabled = False
-        self.btn_scegli_cliente.enabled = False
-        self.btn_agg_cliente.enabled = False
-        self.btn_canc_cliente.enabled = False
-    
-    def agg_cliente(self):
-        dialog = DialogAggCliente()
-        if dialog.exec_() != 0:
-            cliente_nuovo = dialog.cliente_nuovo
-            c = model_to_dict(Cliente.create(**cliente_nuovo))
-            num_riga = self.tab_clienti.rowCount()
-            self.tab_clienti.insertRow(num_riga) 
-            for num_colonna, dato in enumerate(c.values()):
-                    if type(dato) is int:
-                        dato = str(dato)
-                        dato = dato.zfill(4)
-                    if dato == None or dato == "":
-                        dato = "N/A"
-                    self.tab_clienti.setItem(num_riga, num_colonna, QtWidgets.QTableWidgetItem(dato))
-    
-    def popup_nessun_cliente(self):
-        popup = QtWidgets.QMessageBox()
-        popup.resize(600, 300)
-        popup.setWindowTitle("Nessun cliente selezionato")
-        popup.setText("Selezionare un cliente dalla tabella per modificare i dati o eliminare la voce.")
-        popup.setIcon(QtWidgets.QMessageBox.Information)
-        x = popup.exec_()
-    
-    def popup_conferma(self):
-        popup = QtWidgets.QMessageBox()
-        popup.resize(600, 300)
-        popup.setWindowTitle("Azione irreversibile")
-        popup.setText("Attenzione, la cancellazione di un cliente non è reversibile. Procedere?")
-        popup.setIcon(QtWidgets.QMessageBox.Warning)
-        popup.setStandardButtons(QtWidgets.QMessageBox.Ok | QtWidgets.QMessageBox.Cancel)
-        popup.setDefaultButton(QtWidgets.QMessageBox.Cancel)
-        popup.buttonClicked.connect(self.show_popup_btn)
-        x = popup.exec_()
-    
-    def popup_salva_prev(self):
-        popup = QtWidgets.QMessageBox()
-        popup.resize(600, 300)
-        popup.setWindowTitle("Azione irreversibile")
-        popup.setText("Attenzione, quest'azione sovrascriverà i dati precedenti. Procedere?")
-        popup.setIcon(QtWidgets.QMessageBox.Warning)
-        popup.setStandardButtons(QtWidgets.QMessageBox.Ok | QtWidgets.QMessageBox.Cancel)
-        popup.setDefaultButton(QtWidgets.QMessageBox.Cancel)
-        popup.buttonClicked.connect(self.show_popup_btn)
-        popup.exec_()
-    
-    def show_popup_btn(self, item):
-        self.bool_conferma = item.text() == "OK"
-    
-    def edit_cliente(self):
-        if self.tab_clienti.currentRow() == -1:
-            self.popup_nessun_cliente()
-        else:
-            num_riga = self.tab_clienti.currentRow()
-            id_cliente_edit = self.tab_clienti.item(self.tab_clienti.currentRow(), 0).text()
-            self.current_cliente = model_to_dict(Cliente.get(Cliente.id == id_cliente_edit))
-            dialog = DialogEditCliente()
-            if dialog.exec_() != 0:
-                cliente_modificato = dialog.cliente_modificato
-                query = Cliente.update(cliente_modificato).where(Cliente.id == id_cliente_edit)
-                query.execute()
-                c = model_to_dict(Cliente.get(Cliente.id == id_cliente_edit))
-                for num_colonna, dato in enumerate(c.values()):
-                    if type(dato) is int:
-                        dato = str(dato)
-                        dato = dato.zfill(4)
-                    if dato == None or dato == "":
-                        dato = "N/A"
-                    self.tab_clienti.setItem(num_riga, num_colonna, QtWidgets.QTableWidgetItem(dato))
-    
-    def canc_cliente(self):
-        if self.tab_clienti.currentRow() == -1:
-            self.popup_nessun_cliente()
-        else:
-            self.bool_conferma = 0
-            self.popup_conferma()
-            if self.bool_conferma:
-                id_cliente_canc = self.tab_clienti.item(self.tab_clienti.currentRow(), 0).text()
-                Cliente.get(Cliente.id == id_cliente_canc).delete_instance()
-                self.tab_clienti.removeRow(self.tab_clienti.currentRow())
-    
-    def deduci_SO(self):
-        try:
-            ultimo_prev = model_to_dict(Preventivo.select()
-                                    .where(Preventivo.anno == self.current_prev["anno"])
-                                    .order_by(Preventivo.so.desc())
-                                    .get())
-        except:
-            ultimo_prev = None
-        if ultimo_prev == None:
-            self.current_prev["so"] = 1
-        else:
-            self.current_prev["so"] = ultimo_prev["so"] + 1
-        self.edit_so.setText(str(self.current_prev["so"]) + "/" + str(self.current_prev["anno"]))
-    
-    def deduci_PO(self):
-        try:
-            ultimo_prev = model_to_dict(Preventivo.select()
-                                    .where(Preventivo.anno == self.current_prev["anno"])
-                                    .order_by(Preventivo.po.desc())
-                                    .get())
-        except:
-            ultimo_prev = None
-        if ultimo_prev == None:
-            self.current_prev["po"] = 1
-        else:
-            self.current_prev["po"] = ultimo_prev["po"] + 1
-        self.edit_po.setText(str(self.current_prev["po"]) + "/" + str(self.current_prev["anno"]))
-    
-    def deduci_data(self):
-        self.current_prev["data"] = date.today().strftime("%Y-%m-%d")
-        self.current_prev["anno"] = date.today().year
-        self.edit_data_prev.setDate(QtCore.QDate.fromString(self.current_prev["data"], 'yyyy-MM-dd'))
-    
-    def scegli_testo_in(self):
-        self.quale_testo = "in"
-        dialog = DialogScegliTesto()
-        if dialog.exec_() != 0:
-            testo_da_passare = dialog.testo_da_passare
-            if len(testo_da_passare) > 50:
-                testo_da_passare = testo_da_passare[:45] + "..." # elisione
-            self.edit_testo_in.setText(testo_da_passare)
-
-    def scegli_testo_fin(self):
-        self.quale_testo = "fin"
-        dialog = DialogScegliTesto()
-        if dialog.exec_() != 0:
-            testo_da_passare = dialog.testo_da_passare
-            if len(testo_da_passare) > 50:
-                testo_da_passare = testo_da_passare[:45] + "..." # elisione
-            self.edit_testo_fin.setText(testo_da_passare)
-
-    def deduci_testo_in(self):
-        try:
-            ultimo_prev = model_to_dict(Preventivo.select()
-                                    .where(Preventivo.anno == self.current_prev["anno"])
-                                    .get())
-        except:
-            ultimo_prev = None
-        if ultimo_prev == None:
-            self.current_prev["testo_in"] = ""
-        else:
-            self.current_prev["testo_in"] = ultimo_prev["testo_in"]
-        self.edit_testo_in.setText(self.current_prev["testo_in"])
-
-    def deduci_testo_fin(self):
-        try:
-            ultimo_prev = model_to_dict(Preventivo.select()
-                                    .where(Preventivo.anno == self.current_prev["anno"])
-                                    .get())
-        except:
-            ultimo_prev = None
-        if ultimo_prev == None:
-            self.current_prev["testo_fin"] = ""
-        else:
-            self.current_prev["testo_fin"] = ultimo_prev["testo_fin"]
-        self.edit_testo_fin.setText(self.current_prev["testo_fin"])
-        
-    def deduci_nick_cliente(self):
-        pass # wip
-
-    def scegli_progetto(self):
-        pass # wip
-
-    def deduci_descr(self):
-        pass # wip
-
-    def crea_prev(self):
-        if self.tab_clienti.currentRow() == -1:
-            self.popup_nessun_cliente()
-        else:
-            id_cliente = self.tab_clienti.item(self.tab_clienti.currentRow(), 0).text()
-            self.current_cliente = model_to_dict(Cliente.get(Cliente.id == id_cliente))
-            self.current_prev["cliente"] = self.current_cliente["id"]
-            self.deduci_SO()
-            self.deduci_PO()
-            self.deduci_data()
-            self.load_prev(self.current_prev, self.current_cliente)
-    
-    def load_prev(self, prev, cliente):
-        self.stackedWidget.setCurrentIndex(2)
-        self.edit_po.setText("/".join((str(prev["po"]), str(prev["anno"]))))
-        self.edit_so.setText("/".join((str(prev["so"]), str(prev["anno"]))))
-        self.edit_data_prev.setDate(QtCore.QDate.fromString(prev["data"], 'yyyy-MM-dd'))
-        self.edit_nick_cliente.setText(prev["nick_cliente"])
-        self.line_prev_nome.setText(" ".join((cliente["nome"], cliente["cognome"])))
-        self.line_prev_via.setText(cliente["via"])
-        self.line_prev_cantiere.setText(cliente["cantiere"])
-        self.line_prev_comune.setText(cliente["comune"])
-        self.line_prev_tel1.setText(cliente["tel1"])
-        self.line_prev_tel2.setText(cliente["tel2"])
-        self.line_prev_email.setText(cliente["email"])
-
-        self.listwidget_inf.clear()
-        self.lista_id_inf = []
-        query = Infisso.select().where(Infisso.prev == prev["id"])
-        if query.dicts() != None:
-            for num_riga, inf in enumerate(query.dicts()):
-                self.lista_id_inf.append(inf["id"])
-                if inf["posizione"] != ("" or None):
-                    posiz = " - %s" %(inf["posizione"].upper())
-                self.listwidget_inf.addItem("SERRAMENTO %d [%s]%s" %(num_riga, inf["codice"], posiz))
-                # aggiunta pixmap e thumbnail a ogni riga
-        self.view_db_progetti()
-
-    def popola_tree(self, nodo_padre, figlio):
-        nodo_figlio_1 = TreeItem(figlio.codice, 12, color = QtGui.QColor(100, 0, 0))
-        nodo_figlio_2 = TreeItem(figlio.descrizione, 12, color = QtGui.QColor(0, 0, 0))
-        nodo_padre.appendRow([nodo_figlio_1, nodo_figlio_2])
-        if figlio.figli.dicts() != None:
-            for nipote in figlio.figli:
-                self.popola_tree(nodo_figlio_1, nipote)
-    
-    def view_db_progetti(self):
-        query = Progetto.select()
-        casi_base = query.where(Progetto.genitore == None)
-        if casi_base.dicts() != None:
-            # creazione del root_node già fatta
-            for caso_base in casi_base:
-                self.popola_tree(self.root_node, caso_base)
-
-    def load_infisso(self):
-        id_infisso = self.lista_id_inf[self.listwidget_inf.currentRow()]
-        self.current_infisso = Infisso.get(Infisso.id == id_infisso)
-        self.edit_cod_prog.setText(self.current_infisso["cod_prog"])
-        self.edit_posiz.setText(self.current_infisso["posizione"])
-        self.edit_descr.setText(self.current_infisso["descrizione"])
-        self.edit_note_descr.setText(self.current_infisso["note"])
-        # impostare foto 2d e 3d
-        self.spin_num_pz.setValue(self.current_infisso["num_pz"])
-        self.spin_lunghezza.setValue(self.current_infisso["lunghezza"])
-        self.spin_altezza.setValue(self.current_infisso["altezza"])
-        self.spin_spessore.setValue(self.current_infisso["spessore"])
-
-
-    def salva_prev(self):
-        self.popup_salva_prev()
-
-    # <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> <> #
-
-
-    def retranslateUi(self, reGI):
-        _translate = QtCore.QCoreApplication.translate
-        reGI.setWindowTitle(_translate("reGI", "reGI 0.3.X"))
-        self.groupBox.setTitle(_translate("reGI", "Preventivi"))
-        self.label_3.setText(_translate("reGI", "Inizia"))
-        self.label_4.setText(_translate("reGI", "Recenti"))
-        self.btn_nuovo_prev.setText(_translate("reGI", "Nuovo preventivo..."))
-        self.btn_apri_prev.setText(_translate("reGI", "Apri preventivo..."))
-        self.groupBox_2.setTitle(_translate("reGI", "Clienti"))
-        self.btn_view_anagrafica.setText(_translate("reGI", "Visualizza anagrafica clienti"))
-        self.btn_sc_agg_cliente.setText(_translate("reGI", "Aggiungi cliente"))
-        self.btn_sc_edit_cliente.setText(_translate("reGI", "Modifica cliente"))
-        self.groupBox_3.setTitle(_translate("reGI", "Progetti"))
-        self.btn_view_modelli.setText(_translate("reGI", "Visualizza progetti"))
-        self.btn_agg_modello.setText(_translate("reGI", "Aggiungi progetto"))
-        self.btn_edit_modello.setText(_translate("reGI", "Modifica progetto"))
-        self.tab_clienti.setSortingEnabled(True)
-        item = self.tab_clienti.horizontalHeaderItem(0)
-        item.setText(_translate("reGI", "ID"))
-        item = self.tab_clienti.horizontalHeaderItem(1)
-        item.setText(_translate("reGI", "Cognome"))
-        item = self.tab_clienti.horizontalHeaderItem(2)
-        item.setText(_translate("reGI", "Nome"))
-        item = self.tab_clienti.horizontalHeaderItem(3)
-        item.setText(_translate("reGI", "Via"))
-        item = self.tab_clienti.horizontalHeaderItem(4)
-        item.setText(_translate("reGI", "Comune"))
-        item = self.tab_clienti.horizontalHeaderItem(5)
-        item.setText(_translate("reGI", "Tel. 1"))
-        item = self.tab_clienti.horizontalHeaderItem(6)
-        item.setText(_translate("reGI", "Tel. 2"))
-        item = self.tab_clienti.horizontalHeaderItem(7)
-        item.setText(_translate("reGI", "Email"))
-        item = self.tab_clienti.horizontalHeaderItem(8)
-        item.setText(_translate("reGI", "Cantiere"))
-        self.label_2.setText(_translate("reGI", "Seleziona un cliente"))
-        self.btn_edit_cliente.setText(_translate("reGI", "Modifica cliente"))
-        self.btn_scegli_cliente.setText(_translate("reGI", "Scegli cliente"))
-        self.btn_agg_cliente.setText(_translate("reGI", "Aggiungi cliente..."))
-        self.btn_canc_cliente.setText(_translate("reGI", "Elimina cliente"))
-        __sortingEnabled = self.listwidget_inf.isSortingEnabled()
-        self.listwidget_inf.setSortingEnabled(False)
-        item = self.listwidget_inf.item(0)
-        item.setText(_translate("reGI", "Serramento 1 [PF2] - CUCINA"))
-        self.listwidget_inf.setSortingEnabled(__sortingEnabled)
-        self.label_23.setText(_translate("reGI", "Codice progetto"))
-        self.label_24.setText(_translate("reGI", "Descrizione"))
-        self.edit_note_descr.setPlaceholderText(_translate("reGI", "Inserisci delle note sul serramento richiesto dal cliente."))
-        self.label_25.setText(_translate("reGI", "Note"))
-        self.btn_scegli_prog.setText(_translate("reGI", "..."))
-        self.edit_descr.setHtml(_translate("reGI", "<!DOCTYPE HTML PUBLIC \"-//W3C//DTD HTML 4.0//EN\" \"http://www.w3.org/TR/REC-html40/strict.dtd\">\n"
-"<html><head><meta name=\"qrichtext\" content=\"1\" /><style type=\"text/css\">\n"
-"p, li { white-space: pre-wrap; }\n"
-"</style></head><body style=\" font-family:\'MS Shell Dlg 2\'; font-size:10pt; font-weight:600; font-style:normal;\">\n"
-"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-weight:400;\">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.</span></p></body></html>"))
-        self.label_26.setText(_translate("reGI", "Stampa:"))
-        self.checkbox_foto_3d.setText(_translate("reGI", "modello 3D"))
-        self.checkbox_foto_2d.setText(_translate("reGI", "modello 2D"))
-        self.label_27.setText(_translate("reGI", "Posizione"))
-        self.label_32.setText(_translate("reGI", "Dimensioni"))
-        self.label_33.setText(_translate("reGI", "Superficie"))
-        self.btn_salva_prev.setText(_translate("reGI", "Salva preventivo"))
-        self.btn_nuovo_serr.setText(_translate("reGI", "Nuovo"))
-        self.btn_dupl_serr.setText(_translate("reGI", "Duplica"))
-        self.btn_canc_serr.setText(_translate("reGI", "Cancella"))
-        self.radio_prezzo_listino.setText(_translate("reGI", "Prezzo da listino"))
-        self.radio_prezzo_custom.setText(_translate("reGI", "Prezzo personalizzato"))
-        self.label_13.setText(_translate("reGI", "x"))
-        self.label_34.setText(_translate("reGI", "x"))
-        self.label_38.setText(_translate("reGI", "%"))
-        self.radio_sconto.setText(_translate("reGI", "Sconto"))
-        self.radio_sconto_percent.setText(_translate("reGI", "Sconto percentuale"))
-        self.label_39.setText(_translate("reGI", "%"))
-        self.label_37.setText(_translate("reGI", "IVA"))
-        self.edit_iva.setText(_translate("reGI", "22"))
-        self.label_35.setText(_translate("reGI", "Importo lordo"))
-        self.label_12.setText(_translate("reGI", "N. pezzi"))
-        self.spin_num_pz.setSuffix(_translate("reGI", " pz."))
-        self.label_9.setText(_translate("reGI", "Lunghezza"))
-        self.label_10.setText(_translate("reGI", "Altezza"))
-        self.label_11.setText(_translate("reGI", "Spessore"))
-        self.spin_lunghezza.setSuffix(_translate("reGI", " mm"))
-        self.spin_altezza.setSuffix(_translate("reGI", " mm"))
-        self.spin_spessore.setSuffix(_translate("reGI", " mm"))
-        self.label_28.setText(_translate("reGI", "Materiale"))
-        self.label_29.setText(_translate("reGI", "Verniciatura"))
-        self.label_30.setText(_translate("reGI", "Variante 3"))
-        self.combo_materiale.setItemText(0, _translate("reGI", "Legno"))
-        self.combo_materiale.setItemText(1, _translate("reGI", "L/A"))
-        self.combo_vernice.setItemText(0, _translate("reGI", "Standard"))
-        self.label_31.setText(_translate("reGI", "Note sulle varianti"))
-        self.edit_note_varianti.setPlaceholderText(_translate("reGI", "Inserisci delle note sul materiale o sulla verniciatura richiesta dal cliente."))
-        self.tab_dettaglio_totale_stampa.setTabText(self.tab_dettaglio_totale_stampa.indexOf(self.scheda_dettaglio), _translate("reGI", "Dettaglio"))
-        self.label_36.setText(_translate("reGI", "Riepilogo serramenti"))
-        self.tableWidget.setSortingEnabled(True)
-        item = self.tableWidget.horizontalHeaderItem(0)
-        item.setText(_translate("reGI", "Serramento"))
-        item = self.tableWidget.horizontalHeaderItem(1)
-        item.setText(_translate("reGI", "Prezzo unitario lordo"))
-        item = self.tableWidget.horizontalHeaderItem(2)
-        item.setText(_translate("reGI", "N. pezzi"))
-        self.tab_dettaglio_totale_stampa.setTabText(self.tab_dettaglio_totale_stampa.indexOf(self.scheda_totale), _translate("reGI", "Totale"))
-        self.label.setText(_translate("reGI", "N. Progressivo"))
-        self.label_5.setText(_translate("reGI", "Data"))
-        self.btn_scegli_testo_in.setText(_translate("reGI", "..."))
-        self.label_15.setText(_translate("reGI", "Testo iniziale"))
-        self.btn_scegli_testo_fin.setText(_translate("reGI", "..."))
-        self.label_16.setText(_translate("reGI", "Testo finale"))
-        self.label_20.setText(_translate("reGI", "Nome alternativo cliente"))
-        self.radio_so.setText(_translate("reGI", "Preventivo"))
-        self.radio_po.setText(_translate("reGI", "Conferma\n"
-"d\'ordine"))
-        self.label_6.setText(_translate("reGI", "Nome o ragione sociale"))
-        self.label_7.setText(_translate("reGI", "Indirizzo"))
-        self.label_8.setText(_translate("reGI", "Indirizzo cantiere"))
-        self.label_14.setText(_translate("reGI", "Città e provincia"))
-        self.label_17.setText(_translate("reGI", "Telefono 1"))
-        self.label_18.setText(_translate("reGI", "Telefono 2"))
-        self.label_19.setText(_translate("reGI", "Email"))
-        self.menuNuovo.setTitle(_translate("reGI", "File"))
-        self.menuNuovo_2.setTitle(_translate("reGI", "Nuovo..."))
-        self.menuInserisci.setTitle(_translate("reGI", "Inserisci"))
-        self.menuVista.setTitle(_translate("reGI", "Vista"))
-        self.menuSeleziona_stile.setTitle(_translate("reGI", "Design"))
-        self.menuPalette.setTitle(_translate("reGI", "Palette"))
-        self.menuAnteprima.setTitle(_translate("reGI", "Anteprima"))
-        self.menuInfo.setTitle(_translate("reGI", "Info"))
-        self.actionVersione.setText(_translate("reGI", "Versione..."))
-        self.actionVerifica_aggiornamenti.setText(_translate("reGI", "Verifica aggiornamenti..."))
-        self.actionApri.setText(_translate("reGI", "Apri..."))
-        self.actionRecente_1.setText(_translate("reGI", "Recente 1"))
-        self.actionRecente_2.setText(_translate("reGI", "Recente 2"))
-        self.actionDa_zero.setText(_translate("reGI", "Preventivo da zero"))
-        self.actionDa_precompilato.setText(_translate("reGI", "Preventivo da precompilato"))
-        self.actionInfisso.setText(_translate("reGI", "Infisso..."))
-        self.actionVisualizza_anteprima_di_stampa.setText(_translate("reGI", "Visualizza anteprima di stampa"))
-        self.actionDefault.setText(_translate("reGI", "Default"))
-        self.actionFusion.setText(_translate("reGI", "Fusion"))
-        self.actionWindows.setText(_translate("reGI", "Windows"))
-        self.actionWindows_Vista.setText(_translate("reGI", "Windows Vista"))
-        self.actionSchermata_iniziale.setText(_translate("reGI", "Schermata iniziale"))
-        self.actionAnagrafica_clienti.setText(_translate("reGI", "Anagrafica clienti"))
-        self.actionLista_progetti.setText(_translate("reGI", "Lista progetti"))
-        self.actionCliente.setText(_translate("reGI", "Cliente"))
-        self.actionProgetto.setText(_translate("reGI", "Progetto"))
-        self.actionServe_aiuto.setText(_translate("reGI", "Serve aiuto?"))
-        self.actionProgetto_2.setText(_translate("reGI", "Progetto..."))
-        self.actionCliente_2.setText(_translate("reGI", "Cliente..."))
-        self.actionEsci.setText(_translate("reGI", "Esci"))
-        self.actionPalette_1.setText(_translate("reGI", "Palette 1"))
-        self.actionPalette_2.setText(_translate("reGI", "Palette 2"))
-
-
-if __name__ == "__main__":
-    import sys
-    app = QtWidgets.QApplication(sys.argv)
-    reGI = QtWidgets.QMainWindow()
-    reGI.setWindowTitle("reGI a0.3.X")
-    reGI.setWindowIcon(QtGui.QIcon("../Risorse/icon_0.png"))
-    ui = Ui_reGI()
-    ui.setupUi(reGI)
-    reGI.show()
-    sys.exit(app.exec_())
-    db.close()
